Refactor: Extract playfield rendering and input logic
Extract rendering and input handling logic from SimplePlayfield into separate modules. This improves code organization and maintainability by separating concerns.
