+++ assets/playfield.js
@@ -1,6 +1,5 @@
 // Playfield gameplay class extracted from the main bundle for reuse across entry points.
 import { createAlephChainRegistry } from '../scripts/features/towers/alephChain.js';
-import { formatGameNumber } from '../scripts/core/formatting.js';
 import { convertMathExpressionToPlainText } from '../scripts/core/mathText.js';
 import {
   playTowerPlacementNotes,
@@ -18,8 +17,6 @@ import {
 } from './towersTab.js';
 import {
   moteGemState,
-  MOTE_GEM_COLLECTION_RADIUS,
-  collectMoteGemsWithinRadius,
   collectMoteGemDrop,
   spawnMoteGemDrop,
   resetActiveMoteGems,
@@ -40,26 +37,24 @@ import {
 import { colorToRgbaString, resolvePaletteColorStops } from '../scripts/features/towers/powderTower.js';
 import { notifyTowerPlaced } from './achievementsTab.js';
 import { metersToPixels, ALPHA_BASE_RADIUS_FACTOR } from './gameUnits.js'; // Allow playfield interactions to convert standardized meters into pixels.
+import { formatCombatNumber } from './playfield/utils/formatting.js';
+import { easeInCubic, easeOutCubic } from './playfield/utils/math.js';
+import * as CanvasRenderer from './playfield/render/CanvasRenderer.js';
 import {
-  ensureAlphaState as ensureAlphaStateHelper,
-  teardownAlphaTower as teardownAlphaTowerHelper,
-  spawnAlphaAttackBurst as spawnAlphaAttackBurstHelper,
+  PLAYFIELD_VIEW_DRAG_THRESHOLD,
+  PLAYFIELD_VIEW_PAN_MARGIN_METERS,
+} from './playfield/constants.js';
+import * as InputController from './playfield/input/InputController.js';
+import * as HudBindings from './playfield/ui/HudBindings.js';
+import * as TowerManager from './playfield/managers/TowerManager.js';
+import {
   updateAlphaBursts as updateAlphaBurstsHelper,
-  drawAlphaBursts as drawAlphaBurstsHelper,
 } from '../scripts/features/towers/alphaTower.js';
 import {
-  ensureBetaState as ensureBetaStateHelper,
-  teardownBetaTower as teardownBetaTowerHelper,
-  spawnBetaAttackBurst as spawnBetaAttackBurstHelper,
   updateBetaBursts as updateBetaBurstsHelper,
-  drawBetaBursts as drawBetaBurstsHelper,
 } from '../scripts/features/towers/betaTower.js';
 import {
-  ensureGammaState as ensureGammaStateHelper,
-  teardownGammaTower as teardownGammaTowerHelper,
-  spawnGammaAttackBurst as spawnGammaAttackBurstHelper,
   updateGammaBursts as updateGammaBurstsHelper,
-  drawGammaBursts as drawGammaBurstsHelper,
 } from '../scripts/features/towers/gammaTower.js';
 import {
   ensureEpsilonState as ensureEpsilonStateHelper,
@@ -67,41 +62,19 @@ import {
   applyEpsilonHit as applyEpsilonHitHelper,
 } from '../scripts/features/towers/epsilonTower.js';
 import {
-  evaluateZetaMetrics as evaluateZetaMetricsHelper,
-  teardownZetaTower as teardownZetaTowerHelper,
-  ensureZetaState as ensureZetaStateHelper,
   updateZetaTower as updateZetaTowerHelper,
   applyZetaDamage as applyZetaDamageHelper,
-  drawZetaPendulums as drawZetaPendulumsHelper,
 } from '../scripts/features/towers/zetaTower.js';
 import {
-  teardownEtaTower as teardownEtaTowerHelper,
-  ensureEtaState as ensureEtaStateHelper,
-  mergeEtaTower as mergeEtaTowerHelper,
   updateEtaTower as updateEtaTowerHelper,
   fireEtaLaser as fireEtaLaserHelper,
   applyEtaDamage as applyEtaDamageHelper,
-  drawEtaOrbits as drawEtaOrbitsHelper,
   ETA_MAX_PRESTIGE_MERGES,
 } from '../scripts/features/towers/etaTower.js';
 import {
-  ensureDeltaState as ensureDeltaStateHelper,
-  configureDeltaBehavior as configureDeltaBehaviorHelper,
-  teardownDeltaTower as teardownDeltaTowerHelper,
-  updateDeltaAnchors as updateDeltaAnchorsHelper,
-  clearTowerManualTarget as clearTowerManualTargetHelper,
-  getTowerManualTarget as getTowerManualTargetHelper,
   deployDeltaSoldier as deployDeltaSoldierHelper,
   updateDeltaTower as updateDeltaTowerHelper,
-  drawDeltaSoldiers as drawDeltaSoldiersHelper,
 } from '../scripts/features/towers/deltaTower.js';
-
-// Minimum pointer distance before the playfield interprets input as a camera drag.
-const PLAYFIELD_VIEW_DRAG_THRESHOLD = 6;
-// Allow the camera to pan beyond the level edges by a fixed 4 meter buffer regardless of zoom.
-const PLAYFIELD_VIEW_PAN_MARGIN_METERS = 4;
-// Scale battlefield mote gem rendering relative to the canvas so drop sizes stay consistent across devices.
-const GEM_MOTE_BASE_RATIO = 0.02;
 
 // Dependency container allows the main module to provide shared helpers without creating circular imports.
 const defaultDependencies = {
@@ -119,81 +92,6 @@ const defaultDependencies = {
   // Allows the playfield to respect the global graphics fidelity toggle.
   isLowGraphicsMode: () => false,
 };
-
-// Normalize projectile color data so beam rendering can rely on palette-aware RGB objects.
-function normalizeProjectileColor(candidate, fallbackPosition = 1) {
-  if (candidate && typeof candidate === 'object' && Number.isFinite(candidate.r) && Number.isFinite(candidate.g) && Number.isFinite(candidate.b)) {
-    return {
-      r: Math.max(0, Math.min(255, Math.round(candidate.r))),
-      g: Math.max(0, Math.min(255, Math.round(candidate.g))),
-      b: Math.max(0, Math.min(255, Math.round(candidate.b))),
-    };
-  }
-  const ratio = Math.max(0, Math.min(1, fallbackPosition));
-  const fallback = samplePaletteGradient(ratio) || { r: 139, g: 247, b: 255 };
-  return {
-    r: Math.max(0, Math.min(255, Math.round(fallback.r))),
-    g: Math.max(0, Math.min(255, Math.round(fallback.g))),
-    b: Math.max(0, Math.min(255, Math.round(fallback.b))),
-  };
-}
-
-// Present combat-facing numbers with trimmed trailing zeros while respecting notation preferences.
-function formatCombatNumber(value) {
-  const label = formatGameNumber(value);
-  if (typeof label !== 'string') {
-    return label;
-  }
-  const trimmedDecimals = label.replace(/(\.\d*?)(0+)(?=(?:\s| ×|$))/g, (match, decimals) => {
-    const stripped = decimals.replace(/0+$/, '');
-    return stripped;
-  });
-  return trimmedDecimals.replace(/\.($|(?=\s)|(?= ×))/g, '');
-}
-
-// Render an additive gradient blob so shared connection motes inherit the α burst glow.
-function drawConnectionMoteGlow(ctx, x, y, radius, color, opacity = 1) {
-  if (!ctx || !color) {
-    return;
-  }
-  const baseRadius = Math.max(1.6, radius || 2.4);
-  const glowRadius = baseRadius * 2.4;
-  const gradient = ctx.createRadialGradient(x, y, 0, x, y, glowRadius);
-  const alpha = Math.max(0, Math.min(1, opacity));
-  gradient.addColorStop(0, `rgba(${color.r}, ${color.g}, ${color.b}, ${alpha})`);
-  gradient.addColorStop(0.45, `rgba(${color.r}, ${color.g}, ${color.b}, ${alpha * 0.45})`);
-  gradient.addColorStop(1, `rgba(${color.r}, ${color.g}, ${color.b}, 0)`);
-  ctx.fillStyle = gradient;
-  ctx.beginPath();
-  ctx.arc(x, y, glowRadius, 0, Math.PI * 2);
-  ctx.fill();
-}
-
-// Cubic easing helpers keep supply motes and swirl launches smooth and consistent with tower bursts.
-const easeInCubic = (value) => {
-  const clamped = Math.max(0, Math.min(1, value));
-  return clamped * clamped * clamped;
-};
-
-const easeOutCubic = (value) => {
-  const clamped = Math.max(0, Math.min(1, value));
-  const inverted = 1 - clamped;
-  return 1 - inverted * inverted * inverted;
-};
-
-// Preload the Mind Gate sprite so the path finale mirrors the Towers tab art.
-const MIND_GATE_SPRITE_URL = 'assets/images/tower-mind-gate.svg';
-const mindGateSprite = new Image();
-mindGateSprite.src = MIND_GATE_SPRITE_URL;
-mindGateSprite.decoding = 'async';
-mindGateSprite.loading = 'eager';
-
-// Preload the Enemy Gate sprite so the spawn origin echoes the Codex depiction.
-const ENEMY_GATE_SPRITE_URL = 'assets/images/enemy-gate.svg';
-const enemyGateSprite = new Image();
-enemyGateSprite.src = ENEMY_GATE_SPRITE_URL;
-enemyGateSprite.decoding = 'async';
-enemyGateSprite.loading = 'eager';
 
 let playfieldDependencies = { ...defaultDependencies };
 
@@ -511,50 +409,16 @@ export class SimplePlayfield {
 
   // Applies a canvas shadow when high graphics fidelity is active.
   applyCanvasShadow(ctx, color, blur) {
-    if (!ctx) {
-      return;
-    }
-    if (this.isLowGraphicsMode()) {
-      ctx.shadowColor = 'rgba(0, 0, 0, 0)';
-      ctx.shadowBlur = 0;
-      return;
-    }
-    ctx.shadowColor = color || 'rgba(0, 0, 0, 0)';
-    ctx.shadowBlur = Number.isFinite(blur) ? blur : 0;
+    return CanvasRenderer.applyCanvasShadow.call(this, ctx, color, blur);
   }
 
   // Clears active canvas shadow configuration regardless of the fidelity mode.
   clearCanvasShadow(ctx) {
-    if (!ctx) {
-      return;
-    }
-    ctx.shadowColor = 'rgba(0, 0, 0, 0)';
-    ctx.shadowBlur = 0;
+    return CanvasRenderer.clearCanvasShadow.call(this, ctx);
   }
 
   registerSlots() {
-    this.slotButtons.forEach((button) => {
-      const slotId = button.dataset.slotId;
-      const x = Number.parseFloat(button.dataset.slotX);
-      const y = Number.parseFloat(button.dataset.slotY);
-      if (!slotId || Number.isNaN(x) || Number.isNaN(y)) {
-        return;
-      }
-      const slot = {
-        id: slotId,
-        button,
-        normalized: { x, y },
-        tower: null,
-      };
-      this.slots.set(slotId, slot);
-      button.addEventListener('click', () => this.handleSlotInteraction(slot));
-      button.addEventListener('keydown', (event) => {
-        if (event.key === 'Enter' || event.key === ' ') {
-          event.preventDefault();
-          this.handleSlotInteraction(slot);
-        }
-      });
-    });
+    return HudBindings.registerSlots.call(this);
   }
 
   setAvailableTowers(towerIds = []) {
@@ -634,132 +498,31 @@ export class SimplePlayfield {
   }
 
   bindStartButton() {
-    if (!this.startButton) {
-      return;
-    }
-    this.startButton.addEventListener('click', () => this.handleStartButton());
+    return HudBindings.bindStartButton.call(this);
   }
 
   bindSpeedButton() {
-    if (!this.speedButton) {
-      return;
-    }
-    this.speedButton.addEventListener('click', () => {
-      if (this.audio) {
-        this.audio.unlock();
-      }
-      if (!this.isInteractiveLevelActive()) {
-        if (this.messageEl) {
-          this.messageEl.textContent =
-            'Enter an interactive level to adjust the simulation speed.';
-        }
-        return;
-      }
-      this.cycleSpeedMultiplier();
-      if (this.audio) {
-        this.audio.playSfx('uiToggle');
-      }
-    });
+    return HudBindings.bindSpeedButton.call(this);
   }
 
   bindAutoAnchorButton() {
-    if (!this.autoAnchorButton) {
-      return;
-    }
-    this.autoAnchorButton.addEventListener('click', () => {
-      if (this.audio) {
-        this.audio.unlock();
-      }
-      if (!this.isInteractiveLevelActive()) {
-        if (this.messageEl) {
-          this.messageEl.textContent =
-            'Enter an interactive level to auto-lattice recommended anchors.';
-        }
-        return;
-      }
-      this.autoAnchorTowers();
-    });
+    return HudBindings.bindAutoAnchorButton.call(this);
   }
 
   bindAutoWaveCheckbox() {
-    if (!this.autoWaveCheckbox) {
-      return;
-    }
-    this.autoWaveCheckbox.checked = this.autoWaveEnabled;
-    this.autoWaveCheckbox.disabled = true;
-    this.autoWaveCheckbox.addEventListener('change', () => {
-      if (!this.autoWaveCheckbox) {
-        return;
-      }
-      this.autoWaveEnabled = this.autoWaveCheckbox.checked;
-      if (!this.levelActive || !this.levelConfig || this.combatActive) {
-        if (!this.autoWaveEnabled) {
-          this.cancelAutoStart();
-        }
-        return;
-      }
-      if (this.autoWaveEnabled) {
-        this.scheduleAutoStart({ delay: this.autoStartLeadTime });
-      } else {
-        this.cancelAutoStart();
-        if (this.messageEl) {
-          this.messageEl.textContent =
-            'Auto-start disabled—commence waves when your lattice is ready.';
-        }
-      }
-    });
+    return HudBindings.bindAutoWaveCheckbox.call(this);
   }
 
   scheduleAutoStart(options = {}) {
-    if (
-      !this.autoWaveEnabled ||
-      !this.levelActive ||
-      !this.levelConfig ||
-      this.combatActive
-    ) {
-      return;
-    }
-    const delay = Number.isFinite(options.delay)
-      ? Math.max(0, options.delay)
-      : this.autoStartLeadTime;
-    this.cancelAutoStart();
-    if (typeof window === 'undefined') {
-      return;
-    }
-    this.autoStartDeadline = Date.now() + delay * 1000;
-    this.autoStartTimer = window.setTimeout(() => {
-      this.autoStartTimer = null;
-      this.tryAutoStart();
-    }, delay * 1000);
+    return HudBindings.scheduleAutoStart.call(this, options);
   }
 
   cancelAutoStart() {
-    if (this.autoStartTimer) {
-      clearTimeout(this.autoStartTimer);
-      this.autoStartTimer = null;
-    }
-    this.autoStartDeadline = 0;
+    return HudBindings.cancelAutoStart.call(this);
   }
 
   tryAutoStart() {
-    if (
-      !this.autoWaveEnabled ||
-      !this.levelActive ||
-      !this.levelConfig ||
-      this.combatActive
-    ) {
-      return;
-    }
-    if (!this.towers.length) {
-      if (this.messageEl) {
-        this.messageEl.textContent =
-          'Awaiting lattice placements—auto-start resumes once towers are in place.';
-      }
-      this.scheduleAutoStart({ delay: 1.5 });
-      return;
-    }
-    this.autoStartDeadline = 0;
-    this.handleStartButton();
+    return HudBindings.tryAutoStart.call(this);
   }
 
   attachResizeObservers() {
@@ -777,40 +540,11 @@ export class SimplePlayfield {
   }
 
   attachCanvasInteractions() {
-    if (!this.canvas) {
-      return;
-    }
-    this.canvas.addEventListener('pointerdown', this.pointerDownHandler);
-    this.canvas.addEventListener('pointermove', this.pointerMoveHandler);
-    this.canvas.addEventListener('pointerleave', this.pointerLeaveHandler);
-    this.canvas.addEventListener('click', this.pointerClickHandler);
-    this.canvas.addEventListener('pointerup', this.pointerUpHandler);
-    this.canvas.addEventListener('pointercancel', this.pointerUpHandler);
-    this.canvas.addEventListener('lostpointercapture', this.pointerUpHandler);
-    this.canvas.addEventListener('wheel', this.wheelHandler, { passive: false });
+    return InputController.attachCanvasInteractions.call(this);
   }
 
   createEnemyTooltip() {
-    if (!this.container || this.enemyTooltip) {
-      return;
-    }
-
-    const tooltip = document.createElement('div');
-    tooltip.className = 'enemy-tooltip';
-
-    const nameEl = document.createElement('div');
-    nameEl.className = 'enemy-tooltip-name';
-
-    const hpEl = document.createElement('div');
-    hpEl.className = 'enemy-tooltip-hp';
-
-    tooltip.append(nameEl, hpEl);
-    tooltip.setAttribute('aria-hidden', 'true');
-
-    this.container.appendChild(tooltip);
-    this.enemyTooltip = tooltip;
-    this.enemyTooltipNameEl = nameEl;
-    this.enemyTooltipHpEl = hpEl;
+    return HudBindings.createEnemyTooltip.call(this);
   }
 
   syncCanvasSize() {
@@ -1418,25 +1152,11 @@ export class SimplePlayfield {
   }
 
   enableSlots() {
-    this.slots.forEach((slot) => {
-      if (slot.button) {
-        slot.button.disabled = false;
-      }
-    });
+    return HudBindings.enableSlots.call(this);
   }
 
   disableSlots(clear = false) {
-    this.slots.forEach((slot) => {
-      if (!slot.button) {
-        return;
-      }
-      slot.button.disabled = true;
-      if (clear) {
-        slot.tower = null;
-        slot.button.classList.remove('tower-built');
-        slot.button.setAttribute('aria-pressed', 'false');
-      }
-    });
+    return HudBindings.disableSlots.call(this, clear);
   }
 
   isInteractiveLevelActive() {
@@ -1468,17 +1188,7 @@ export class SimplePlayfield {
   }
 
   updateSpeedButton() {
-    if (!this.speedButton) {
-      return;
-    }
-    const label = this.formatSpeedMultiplier(this.speedMultiplier);
-    this.speedButton.textContent = `Speed ×${label}`;
-    const interactive = this.isInteractiveLevelActive();
-    this.speedButton.disabled = !interactive;
-    this.speedButton.setAttribute('aria-disabled', interactive ? 'false' : 'true');
-    this.speedButton.title = interactive
-      ? 'Cycle the manual defense speed multiplier.'
-      : 'Simulation speed adjusts during the interactive defense.';
+    return HudBindings.updateSpeedButton.call(this);
   }
 
   getAutoAnchorStatus() {
@@ -1504,71 +1214,11 @@ export class SimplePlayfield {
   }
 
   updateAutoAnchorButton() {
-    if (!this.autoAnchorButton) {
-      return;
-    }
-
-    this.autoAnchorButton.textContent = 'Loadout Placement';
-    this.autoAnchorButton.disabled = true;
-    this.autoAnchorButton.setAttribute('aria-disabled', 'true');
-    this.autoAnchorButton.title = 'Drag towers from the loadout to lattice them on the field.';
+    return HudBindings.updateAutoAnchorButton.call(this);
   }
 
   autoAnchorTowers() {
-    if (!this.isInteractiveLevelActive()) {
-      if (this.audio) {
-        this.audio.playSfx('error');
-      }
-      return;
-    }
-    const anchors = Array.isArray(this.levelConfig?.autoAnchors)
-      ? this.levelConfig.autoAnchors
-      : [];
-    if (!anchors.length) {
-      if (this.messageEl) {
-        this.messageEl.textContent = 'No auto-lattice anchors configured for this level yet.';
-      }
-      return;
-    }
-
-    const tolerance = this.anchorTolerance;
-    let placed = 0;
-    let insufficientEnergy = false;
-
-    for (const anchor of anchors) {
-      const occupied = this.towers.some((tower) => {
-        const dx = tower.normalized.x - anchor.x;
-        const dy = tower.normalized.y - anchor.y;
-        return Math.hypot(dx, dy) <= tolerance;
-      });
-      if (occupied) {
-        continue;
-      }
-      if (this.energy < this.levelConfig.towerCost) {
-        insufficientEnergy = true;
-        break;
-      }
-      const success = this.addTowerAt(anchor, { silent: true });
-      if (success) {
-        placed += 1;
-      }
-    }
-
-    const { total, placed: nowPlaced } = this.getAutoAnchorStatus();
-    const remaining = Math.max(0, total - nowPlaced);
-
-    if (typeof this.dependencies.notifyAutoAnchorUsed === 'function') {
-      this.dependencies.notifyAutoAnchorUsed(nowPlaced, total);
-    }
-
-    if (this.audio && placed > 0) {
-      this.audio.playSfx('towerPlace');
-      playTowerPlacementNotes(this.audio, placed);
-    }
-
-    if (this.messageEl) {
-      this.messageEl.textContent = 'Auto-lattice is disabled—drag towers from the loadout instead.';
-    }
+    return HudBindings.autoAnchorTowers.call(this);
   }
 
   canRetryCurrentWave() {
@@ -1693,324 +1343,19 @@ export class SimplePlayfield {
   }
 
   handleCanvasPointerMove(event) {
-    if (!this.levelActive || !this.levelConfig) {
-      this.clearPlacementPreview();
-      this.pointerPosition = null;
-      this.clearEnemyHover();
-      return;
-    }
-
-    const isTouchPointer = event.pointerType === 'touch';
-    // Track touch samples so the same movement logic supports both pinch and one-finger pans.
-    if (isTouchPointer) {
-      this.activePointers.set(event.pointerId, { clientX: event.clientX, clientY: event.clientY });
-      if (this.activePointers.size >= 2) {
-        if (typeof event.preventDefault === 'function') {
-          event.preventDefault();
-        }
-        this.performPinchZoom();
-        this.pointerPosition = null;
-        this.clearPlacementPreview();
-        this.clearEnemyHover();
-        return;
-      }
-    } else {
-      this.activePointers.clear();
-      this.pinchState = null;
-      this.isPinchZooming = false;
-    }
-
-    if (this.connectionDragState.pointerId === event.pointerId) {
-      if (typeof event.preventDefault === 'function') {
-        event.preventDefault();
-      }
-      const normalized = this.getNormalizedFromEvent(event);
-      if (!normalized) {
-        this.connectionDragState.currentPosition = null;
-        this.connectionDragState.hoverEntry = null;
-        if (!this.shouldAnimate) {
-          this.draw();
-        }
-        return;
-      }
-      const position = this.getCanvasPosition(normalized);
-      this.connectionDragState.currentPosition = position ? { ...position } : null;
-      if (this.connectionDragState.startPosition && !this.connectionDragState.active && position) {
-        const dx = position.x - this.connectionDragState.startPosition.x;
-        const dy = position.y - this.connectionDragState.startPosition.y;
-        const distance = Math.hypot(dx, dy);
-        if (Number.isFinite(distance) && distance >= PLAYFIELD_VIEW_DRAG_THRESHOLD) {
-          this.connectionDragState.active = true;
-          this.connectionDragState.hasMoved = true;
-          this.suppressNextCanvasClick = true;
-        }
-      }
-      if (this.connectionDragState.active && position) {
-        this.updateConnectionDragHighlights(position);
-      }
-      if (!this.shouldAnimate) {
-        this.draw();
-      }
-      return;
-    }
-
-    const isPanPointer =
-      this.viewDragState.pointerId !== null &&
-      event.pointerId === this.viewDragState.pointerId &&
-      (!isTouchPointer || this.activePointers.size < 2);
-    if (isPanPointer) {
-      // Translate pointer movement into a clamped camera offset while dragging.
-      const dx = event.clientX - this.viewDragState.startX;
-      const dy = event.clientY - this.viewDragState.startY;
-      const dragDistance = Math.hypot(dx, dy);
-      if (!this.viewDragState.isDragging && dragDistance >= PLAYFIELD_VIEW_DRAG_THRESHOLD) {
-        this.viewDragState.isDragging = true;
-        this.suppressNextCanvasClick = true;
-      }
-      if (this.viewDragState.isDragging) {
-        if (typeof event.preventDefault === 'function') {
-          event.preventDefault();
-        }
-        this.pointerPosition = null;
-        this.clearPlacementPreview();
-        this.clearEnemyHover();
-        const width = this.renderWidth || (this.canvas ? this.canvas.clientWidth : 0) || 1;
-        const height = this.renderHeight || (this.canvas ? this.canvas.clientHeight : 0) || 1;
-        const scale = this.viewScale || 1;
-        const startCenter = this.viewDragState.startCenter || { x: 0.5, y: 0.5 };
-        const nextCenter = {
-          x: startCenter.x - dx / (scale * width),
-          y: startCenter.y - dy / (scale * height),
-        };
-        this.viewCenterNormalized = this.clampViewCenterNormalized(nextCenter);
-        this.applyViewConstraints();
-        this.draw();
-        return;
-      }
-    }
-
-    const normalized = this.getNormalizedFromEvent(event);
-    if (!normalized) {
-      this.clearPlacementPreview();
-      this.pointerPosition = null;
-      this.clearEnemyHover();
-      return;
-    }
-
-    this.pointerPosition = normalized;
-    const position = this.getCanvasPosition(normalized);
-    const hoveredEnemy = this.findEnemyAt(position);
-    if (hoveredEnemy) {
-      this.setEnemyHover(hoveredEnemy.enemy);
-    } else {
-      this.clearEnemyHover();
-    }
-    const hoveredTower = this.findTowerAt(position);
-    if (!this.draggingTowerType && hoveredTower) {
-      this.hoverPlacement = {
-        normalized: { ...hoveredTower.normalized },
-        position: { x: hoveredTower.x, y: hoveredTower.y },
-        range: hoveredTower.range,
-        valid: false,
-        target: hoveredTower,
-        towerType: hoveredTower.type,
-        reason: 'Open lattice menu.',
-      };
-      if (!this.shouldAnimate) {
-        this.draw();
-      }
-      return;
-    }
-
-    const activeType = this.draggingTowerType;
-    if (activeType) {
-      this.updatePlacementPreview(normalized, {
-        towerType: activeType,
-        dragging: Boolean(this.draggingTowerType),
-      });
-    } else {
-      this.clearPlacementPreview();
-    }
-
-    if (!this.shouldAnimate) {
-      this.draw();
-    }
+    return InputController.handleCanvasPointerMove.call(this, event);
   }
 
   performPinchZoom() {
-    const pointers = Array.from(this.activePointers.values());
-    if (pointers.length < 2) {
-      this.pinchState = null;
-      this.isPinchZooming = false;
-      return;
-    }
-    const [first, second] = pointers;
-    const dx = first.clientX - second.clientX;
-    const dy = first.clientY - second.clientY;
-    const distance = Math.hypot(dx, dy);
-    if (!Number.isFinite(distance) || distance <= 0) {
-      return;
-    }
-    const midpoint = {
-      clientX: (first.clientX + second.clientX) / 2,
-      clientY: (first.clientY + second.clientY) / 2,
-    };
-    const anchor = this.getCanvasRelativeFromClient(midpoint);
-    if (!anchor) {
-      return;
-    }
-    this.isPinchZooming = true;
-    if (!this.pinchState || !Number.isFinite(this.pinchState.startDistance) || this.pinchState.startDistance <= 0) {
-      this.pinchState = {
-        startDistance: distance,
-        startScale: this.viewScale,
-      };
-      return;
-    }
-    const baseDistance = this.pinchState.startDistance;
-    const baseScale = this.pinchState.startScale || this.viewScale;
-    if (!Number.isFinite(baseDistance) || baseDistance <= 0) {
-      return;
-    }
-    const targetScale = (distance / baseDistance) * baseScale;
-    this.setZoom(targetScale, anchor);
-    this.pinchState.startScale = this.viewScale;
-    this.pinchState.startDistance = distance;
+    return InputController.performPinchZoom.call(this);
   }
 
   handleCanvasPointerDown(event) {
-    const isTouchPointer = event.pointerType === 'touch';
-    if (isTouchPointer) {
-      this.activePointers.set(event.pointerId, { clientX: event.clientX, clientY: event.clientY });
-      if (this.activePointers.size < 2) {
-        this.pinchState = null;
-        this.isPinchZooming = false;
-      } else {
-        // Cancel any in-progress pan when a second touch begins a pinch gesture.
-        this.viewDragState.pointerId = null;
-        this.viewDragState.isDragging = false;
-        this.suppressNextCanvasClick = true;
-      }
-    } else {
-      this.activePointers.clear();
-      this.pinchState = null;
-      this.isPinchZooming = false;
-    }
-
-    if (!this.levelConfig) {
-      return;
-    }
-
-    const isPrimaryMouseDrag = !isTouchPointer && event.button === 0;
-    const isSingleTouchDrag = isTouchPointer && this.activePointers.size < 2;
-    const canInitiateDrag = !this.draggingTowerType && (isPrimaryMouseDrag || isSingleTouchDrag);
-    if (canInitiateDrag) {
-      const normalized = this.getNormalizedFromEvent(event);
-      const position = normalized ? this.getCanvasPosition(normalized) : null;
-      const tower = position ? this.findTowerAt(position) : null;
-      if (tower && (tower.type === 'alpha' || tower.type === 'beta' || tower.type === 'gamma')) {
-        this.clearConnectionDragState();
-        this.connectionDragState.pointerId = event.pointerId;
-        this.connectionDragState.originTowerId = tower.id;
-        this.connectionDragState.startPosition = position ? { ...position } : null;
-        this.connectionDragState.currentPosition = position ? { ...position } : null;
-        this.connectionDragState.highlightEntries = [];
-        this.connectionDragState.hoverEntry = null;
-        this.suppressNextCanvasClick = false;
-        if (!isTouchPointer && typeof this.canvas?.setPointerCapture === 'function') {
-          try {
-            this.canvas.setPointerCapture(event.pointerId);
-          } catch (error) {
-            // Ignore pointer capture errors—connection drags still function without it.
-          }
-        }
-        return;
-      }
-    }
-
-    if (canInitiateDrag) {
-      // Allow both mouse input and single-finger touches to initiate camera panning.
-      this.viewDragState.pointerId = event.pointerId;
-      this.viewDragState.startX = event.clientX;
-      this.viewDragState.startY = event.clientY;
-      this.viewDragState.startCenter = { ...(this.viewCenterNormalized || { x: 0.5, y: 0.5 }) };
-      this.viewDragState.isDragging = false;
-      this.suppressNextCanvasClick = false;
-      if (!isTouchPointer && typeof this.canvas?.setPointerCapture === 'function') {
-        try {
-          this.canvas.setPointerCapture(event.pointerId);
-        } catch (error) {
-          // Ignore pointer capture failures so mouse dragging still functions.
-        }
-      }
-    }
+    return InputController.handleCanvasPointerDown.call(this, event);
   }
 
   handleCanvasPointerUp(event) {
-    const isTouchPointer = event.pointerType === 'touch';
-    if (isTouchPointer) {
-      this.activePointers.delete(event.pointerId);
-      if (this.activePointers.size < 2) {
-        this.pinchState = null;
-        this.isPinchZooming = false;
-      }
-    } else {
-      this.activePointers.clear();
-      this.pinchState = null;
-      this.isPinchZooming = false;
-    }
-
-    if (this.connectionDragState.pointerId === event.pointerId) {
-      const dragState = this.connectionDragState;
-      const entry = dragState.hoverEntry;
-      if (dragState.active && entry) {
-        if (entry.action === 'connect') {
-          const connected = this.addTowerConnection(entry.sourceId, entry.targetId);
-          if (connected) {
-            this.suppressNextCanvasClick = true;
-          }
-        } else if (entry.action === 'disconnect') {
-          const removed = this.removeTowerConnection(entry.sourceId, entry.targetId);
-          if (removed) {
-            this.suppressNextCanvasClick = true;
-          }
-        }
-        if (!this.shouldAnimate) {
-          this.draw();
-        }
-      } else if (dragState.hasMoved) {
-        this.suppressNextCanvasClick = true;
-        if (!this.shouldAnimate) {
-          this.draw();
-        }
-      }
-      this.clearConnectionDragState();
-      if (!isTouchPointer && typeof this.canvas?.releasePointerCapture === 'function') {
-        try {
-          this.canvas.releasePointerCapture(event.pointerId);
-        } catch (error) {
-          // Ignore browsers that throw if the pointer was not previously captured.
-        }
-      }
-      return;
-    }
-
-    if (this.viewDragState.pointerId === event.pointerId) {
-      if (this.viewDragState.isDragging) {
-        this.suppressNextCanvasClick = true;
-      }
-      this.viewDragState.pointerId = null;
-      this.viewDragState.isDragging = false;
-    }
-
-    if (!isTouchPointer && typeof this.canvas?.releasePointerCapture === 'function') {
-      try {
-        // Release mouse pointer capture so future drags start from a clean state.
-        this.canvas.releasePointerCapture(event.pointerId);
-      } catch (error) {
-        // Ignore browsers that throw if the pointer was not previously captured.
-      }
-    }
+    return InputController.handleCanvasPointerUp.call(this, event);
   }
 
   updatePlacementPreview(normalized, options = {}) {
@@ -2105,209 +1450,40 @@ export class SimplePlayfield {
   }
 
   handleCanvasPointerLeave() {
-    this.pointerPosition = null;
-    this.clearPlacementPreview();
-    this.clearEnemyHover();
-    this.activePointers.clear();
-    this.pinchState = null;
-    this.isPinchZooming = false;
-    this.viewDragState.pointerId = null;
-    this.viewDragState.isDragging = false;
-    this.clearConnectionDragState();
+    return InputController.handleCanvasPointerLeave.call(this);
   }
 
   // Attempt to gather any mote gems located near the pointer position.
   collectMoteGemsNear(position) {
-    if (!position) {
-      return 0;
-    }
-    const collected = collectMoteGemsWithinRadius(position, MOTE_GEM_COLLECTION_RADIUS, {
-      reason: 'manual',
-    });
-    if (collected > 0 && this.audio) {
-      this.audio.playSfx('uiConfirm');
-    }
-    return collected;
+    return InputController.collectMoteGemsNear.call(this, position);
   }
 
   handleCanvasClick(event) {
-    if (this.audio) {
-      this.audio.unlock();
-    }
-    if (!this.levelActive || !this.levelConfig) {
-      return;
-    }
-
-    if (this.suppressNextCanvasClick) {
-      this.suppressNextCanvasClick = false;
-      return;
-    }
-
-    if (this.isPinchZooming) {
-      return;
-    }
-
-    if (typeof event.preventDefault === 'function') {
-      event.preventDefault();
-    }
-
-    const normalized = this.getNormalizedFromEvent(event);
-    if (!normalized) {
-      return;
-    }
-
-    const position = this.getCanvasPosition(normalized);
-
-    if (
-      typeof this.dependencies.handleDeveloperMapPlacement === 'function' &&
-      this.dependencies.handleDeveloperMapPlacement({
-        normalized,
-        position,
-        playfield: this,
-        event,
-      })
-    ) {
-      return;
-    }
-
-    const enemyTarget = this.findEnemyAt(position);
-    const menuTower = this.getActiveMenuTower();
-    if (enemyTarget) {
-      if (menuTower && this.handleTowerMenuEnemySelection(menuTower, enemyTarget.enemy)) {
-        return;
-      }
-      this.toggleEnemyFocus(enemyTarget.enemy);
-      return;
-    }
-
-    const crystalTarget = this.findCrystalAt(position);
-    if (crystalTarget) {
-      this.toggleCrystalFocus(crystalTarget);
-      return;
-    }
-
-    if (this.activeTowerMenu && this.handleTowerMenuClick(position)) {
-      return;
-    }
-
-    if (this.collectMoteGemsNear(position)) {
-      return;
-    }
-
-    const tower = this.findTowerAt(position);
-    if (tower) {
-      if (this.activeTowerMenu?.towerId === tower.id) {
-        this.closeTowerMenu();
-      } else {
-        this.openTowerMenu(tower);
-      }
-      return;
-    }
-
-    if (this.activeTowerMenu) {
-      this.closeTowerMenu();
-    }
+    return InputController.handleCanvasClick.call(this, event);
   }
 
   handleCanvasWheel(event) {
-    if (!this.levelActive || !this.levelConfig) {
-      return;
-    }
-    if (typeof event.preventDefault === 'function') {
-      event.preventDefault();
-    }
-    if (typeof this.dependencies.isFieldNotesOverlayVisible === 'function' && this.dependencies.isFieldNotesOverlayVisible()) {
-      return;
-    }
-    const anchor = this.getCanvasRelativeFromClient({ clientX: event.clientX, clientY: event.clientY });
-    if (!anchor) {
-      return;
-    }
-    const delta = Number.isFinite(event.deltaY) ? event.deltaY : 0;
-    const factor = Math.exp(-delta * 0.0015);
-    this.applyZoomFactor(factor, anchor);
+    return InputController.handleCanvasWheel.call(this, event);
   }
 
   applyZoomFactor(factor, anchor) {
-    if (!Number.isFinite(factor) || factor <= 0) {
-      return;
-    }
-    const targetScale = this.viewScale * factor;
-    this.setZoom(targetScale, anchor);
+    return InputController.applyZoomFactor.call(this, factor, anchor);
   }
 
   setZoom(targetScale, anchor) {
-    if (!Number.isFinite(targetScale)) {
-      return false;
-    }
-    const width = this.renderWidth || (this.canvas ? this.canvas.clientWidth : 0) || 0;
-    const height = this.renderHeight || (this.canvas ? this.canvas.clientHeight : 0) || 0;
-    const screenPoint = anchor || { x: width / 2, y: height / 2 };
-    const anchorWorld = this.screenToWorld(screenPoint);
-    const clampedScale = Math.min(this.maxViewScale, Math.max(this.minViewScale, targetScale));
-    const previousScale = this.viewScale;
-    this.viewScale = clampedScale;
-    this.applyViewConstraints();
-    if (anchorWorld) {
-      const offsetX = (screenPoint.x - width / 2) / this.viewScale;
-      const offsetY = (screenPoint.y - height / 2) / this.viewScale;
-      this.setViewCenterFromWorld({
-        x: anchorWorld.x - offsetX,
-        y: anchorWorld.y - offsetY,
-      });
-    } else {
-      this.applyViewConstraints();
-    }
-    this.applyViewConstraints();
-    const scaleChanged = Math.abs(previousScale - this.viewScale) > 0.0001;
-    this.draw();
-    return scaleChanged;
+    return InputController.setZoom.call(this, targetScale, anchor);
   }
 
   clearPlacementPreview() {
-    if (!this.hoverPlacement) {
-      return;
-    }
-    this.hoverPlacement = null;
-    if (!this.shouldAnimate) {
-      this.draw();
-    }
+    return InputController.clearPlacementPreview.call(this);
   }
 
   clearEnemyHover() {
-    this.hoverEnemy = null;
-    if (this.enemyTooltip) {
-      this.enemyTooltip.dataset.visible = 'false';
-      this.enemyTooltip.setAttribute('aria-hidden', 'true');
-    }
+    return InputController.clearEnemyHover.call(this);
   }
 
   getNormalizedFromEvent(event) {
-    if (!this.canvas) {
-      return null;
-    }
-    const rect = this.canvas.getBoundingClientRect();
-    const width = rect.width || this.renderWidth || 0;
-    const height = rect.height || this.renderHeight || 0;
-    if (!width || !height) {
-      return null;
-    }
-    const relative = {
-      x: event.clientX - rect.left,
-      y: event.clientY - rect.top,
-    };
-    if (!Number.isFinite(relative.x) || !Number.isFinite(relative.y)) {
-      return null;
-    }
-    const world = this.screenToWorld(relative);
-    if (!world) {
-      return null;
-    }
-    const normalized = {
-      x: world.x / (this.renderWidth || width),
-      y: world.y / (this.renderHeight || height),
-    };
-    return this.clampNormalized(normalized);
+    return InputController.getNormalizedFromEvent.call(this, event);
   }
 
   findTowerAt(position) {
@@ -2597,194 +1773,151 @@ export class SimplePlayfield {
    * upgrade overlay without duplicating the underlying formulas.
    */
   evaluateZetaMetrics(tower) {
-    return evaluateZetaMetricsHelper(this, tower);
+    return TowerManager.evaluateZetaMetrics.call(this, tower);
   }
 
   /**
    * Clear cached α particle data when the tower departs or retunes.
    */
   teardownAlphaTower(tower) {
-    teardownAlphaTowerHelper(this, tower);
+    return TowerManager.teardownAlphaTower.call(this, tower);
   }
 
   /**
    * Ensure α towers keep their particle calibration in sync with canvas scale.
    */
   ensureAlphaState(tower) {
-    return ensureAlphaStateHelper(this, tower);
+    return TowerManager.ensureAlphaState.call(this, tower);
   }
 
   /**
    * Emit α particle swarms whenever the lattice releases an attack.
    */
   spawnAlphaAttackBurst(tower, targetInfo, options = {}) {
-    return spawnAlphaAttackBurstHelper(this, tower, targetInfo, options);
+    return TowerManager.spawnAlphaAttackBurst.call(this, tower, targetInfo, options);
   }
 
   /**
    * Clear cached β particle data when the lattice retunes or is removed.
    */
   teardownBetaTower(tower) {
-    teardownBetaTowerHelper(this, tower);
+    return TowerManager.teardownBetaTower.call(this, tower);
   }
 
   /**
    * Keep β tower particle scaling aligned with the active canvas resolution.
    */
   ensureBetaState(tower) {
-    return ensureBetaStateHelper(this, tower);
+    return TowerManager.ensureBetaState.call(this, tower);
   }
 
   /**
    * Emit β particle bursts using the shared lattice particle system.
    */
   spawnBetaAttackBurst(tower, targetInfo, options = {}) {
-    return spawnBetaAttackBurstHelper(this, tower, targetInfo, options);
+    return TowerManager.spawnBetaAttackBurst.call(this, tower, targetInfo, options);
   }
 
   /**
    * Clear cached γ particle data when the lattice retunes or leaves the field.
    */
   teardownGammaTower(tower) {
-    teardownGammaTowerHelper(this, tower);
+    return TowerManager.teardownGammaTower.call(this, tower);
   }
 
   /**
    * Maintain γ particle scaling so laser visuals stay anchored to the field.
    */
   ensureGammaState(tower) {
-    return ensureGammaStateHelper(this, tower);
+    return TowerManager.ensureGammaState.call(this, tower);
   }
 
   /**
    * Emit γ piercing laser bursts that reuse the shared particle animation stack.
    */
   spawnGammaAttackBurst(tower, targetInfo, options = {}) {
-    return spawnGammaAttackBurstHelper(this, tower, targetInfo, options);
+    return TowerManager.spawnGammaAttackBurst.call(this, tower, targetInfo, options);
   }
 
   /**
    * Clear cached ζ pendulum data when the tower is removed or retuned.
    */
   teardownZetaTower(tower) {
-    teardownZetaTowerHelper(this, tower);
+    return TowerManager.teardownZetaTower.call(this, tower);
   }
 
   /**
    * Ensure ζ towers maintain their double-pendulum state and derived stats.
    */
   ensureZetaState(tower) {
-    return ensureZetaStateHelper(this, tower);
+    return TowerManager.ensureZetaState.call(this, tower);
   }
 
   /**
    * Clear cached η orbital data when the tower is removed or retuned.
    */
   teardownEtaTower(tower) {
-    teardownEtaTowerHelper(this, tower);
+    return TowerManager.teardownEtaTower.call(this, tower);
   }
 
   /**
    * Ensure η towers maintain their orbital ring state and derived stats.
    */
   ensureEtaState(tower, options = {}) {
-    return ensureEtaStateHelper(this, tower, options);
+    return TowerManager.ensureEtaState.call(this, tower, options);
   }
 
   /**
    * Merge η towers to unfold rings and handle prestige transitions.
    */
   mergeEtaTower(tower, { silent = false } = {}) {
-    return mergeEtaTowerHelper(this, tower, { silent });
+    return TowerManager.mergeEtaTower.call(this, tower, { silent });
   }
 
   applyTowerBehaviorDefaults(tower) {
-    if (!tower) {
-      return;
-    }
-    if (!tower.targetPriority) {
-      tower.targetPriority = 'first';
-    }
-    if (!tower.behaviorMode) {
-      tower.behaviorMode = 'pursuit';
-    }
-    if (tower.type === 'alpha') {
-      this.ensureAlphaState(tower);
-    } else if (tower.alphaState) {
-      this.teardownAlphaTower(tower);
-    }
-    if (tower.type === 'beta') {
-      this.ensureBetaState(tower);
-    } else if (tower.betaState) {
-      this.teardownBetaTower(tower);
-    }
-    if (tower.type === 'gamma') {
-      this.ensureGammaState(tower);
-    } else if (tower.gammaState) {
-      this.teardownGammaTower(tower);
-    }
-    if (tower.type === 'delta') {
-      this.ensureDeltaState(tower);
-      this.configureDeltaBehavior(tower, tower.behaviorMode);
-    } else if (tower.deltaState) {
-      this.teardownDeltaTower(tower);
-    }
-    if (tower.type === 'zeta') {
-      // Activate ζ pendulum state so orbit physics stay ready for combat or idle motion.
-      this.ensureZetaState(tower);
-    } else if (tower.zetaState) {
-      // Clean up ζ caches if the lattice retunes into another form.
-      this.teardownZetaTower(tower);
-    }
-    if (tower.type === 'eta') {
-      // Maintain η orbital state so rings stay synchronized while idle.
-      this.ensureEtaState(tower);
-    } else if (tower.etaState) {
-      // Clear η caches when the lattice shifts into another configuration.
-      this.teardownEtaTower(tower);
-    }
+    return TowerManager.applyTowerBehaviorDefaults.call(this, tower);
   }
 
   /**
    * Create or update the Delta state container so soldiers can be simulated.
    */
   ensureDeltaState(tower) {
-    return ensureDeltaStateHelper(this, tower);
+    return TowerManager.ensureDeltaState.call(this, tower);
   }
 
   /**
    * Update Delta soldier stance and anchor points when the player changes modes.
    */
   configureDeltaBehavior(tower, mode) {
-    configureDeltaBehaviorHelper(this, tower, mode);
+    return TowerManager.configureDeltaBehavior.call(this, tower, mode);
   }
 
   /**
    * Remove any lingering Delta data once a tower is dismantled or upgraded away.
    */
   teardownDeltaTower(tower) {
-    teardownDeltaTowerHelper(this, tower);
+    return TowerManager.teardownDeltaTower.call(this, tower);
   }
 
   /**
    * Keep the cached track-hold anchor aligned with the latest path layout.
    */
   updateDeltaAnchors(tower) {
-    updateDeltaAnchorsHelper(this, tower);
+    return TowerManager.updateDeltaAnchors.call(this, tower);
   }
 
   /**
    * Clear any manual Delta target.
    */
   clearTowerManualTarget(tower) {
-    return clearTowerManualTargetHelper(this, tower);
+    return TowerManager.clearTowerManualTarget.call(this, tower);
   }
 
   /**
    * Retrieve the live enemy chosen for manual Delta focus.
    */
   getTowerManualTarget(tower) {
-    return getTowerManualTargetHelper(this, tower);
+    return TowerManager.getTowerManualTarget.call(this, tower);
   }
   getEnemyVisualMetrics(enemy) {
     if (!enemy) {
@@ -4071,66 +3204,14 @@ export class SimplePlayfield {
    * Render swirling supply motes around a lattice.
    */
   drawTowerConnectionParticles(ctx, tower, bodyRadius) {
-    if (!ctx || !tower) {
-      return;
-    }
-    const particles = Array.isArray(tower.connectionParticles) ? tower.connectionParticles : [];
-    if (!particles.length) {
-      return;
-    }
-    ctx.save();
-    ctx.globalCompositeOperation = 'lighter'; // Blend motes additively to mirror α burst luminosity.
-    particles.forEach((particle) => {
-      if (!particle || particle.state === 'done') {
-        return;
-      }
-      const baseColor = particle.type === 'beta'
-        ? { r: 255, g: 214, b: 112 }
-        : { r: 255, g: 138, b: 216 };
-      const color = normalizeProjectileColor(baseColor, 1);
-      const size = particle.size || 2.6;
-      let position = null;
-      if (particle.state === 'launch' || particle.state === 'arrive') {
-        position = particle.position || this.resolveConnectionOrbitAnchor(tower, particle);
-      } else {
-        position = this.resolveConnectionOrbitPosition(tower, particle, bodyRadius);
-      }
-      if (!position) {
-        return;
-      }
-      drawConnectionMoteGlow(ctx, position.x, position.y, size, color, particle.state === 'launch' ? 0.9 : 0.85);
-    });
-    ctx.restore();
+    return CanvasRenderer.drawTowerConnectionParticles.call(this, ctx, tower, bodyRadius);
   }
 
   /**
    * Render flowing motes along each established connection link.
    */
   drawConnectionEffects(ctx) {
-    if (!ctx || !Array.isArray(this.connectionEffects) || !this.connectionEffects.length) {
-      return;
-    }
-    ctx.save();
-    ctx.globalCompositeOperation = 'lighter'; // Keep link motes luminous like α burst particles.
-    this.connectionEffects.forEach((effect) => {
-      const source = effect.source || this.getTowerById(effect.sourceId);
-      const target = effect.target || this.getTowerById(effect.targetId);
-      if (!source || !target) {
-        return;
-      }
-      const baseColor = source.type === 'beta'
-        ? { r: 255, g: 214, b: 112 }
-        : { r: 255, g: 138, b: 216 };
-      const color = normalizeProjectileColor(baseColor, 1);
-      effect.particles.forEach((particle) => {
-        const progress = Math.max(0, Math.min(1, particle.progress || 0));
-        const x = source.x + (target.x - source.x) * progress;
-        const y = source.y + (target.y - source.y) * progress;
-        // Render the travelling mote with the shared glow helper for consistency.
-        drawConnectionMoteGlow(ctx, x, y, 3.2, color, 0.7);
-      });
-    });
-    ctx.restore();
+    return CanvasRenderer.drawConnectionEffects.call(this, ctx);
   }
 
   validatePlacement(normalized, options = {}) {
@@ -6095,99 +5176,11 @@ export class SimplePlayfield {
   }
 
   updateHud() {
-    if (this.waveEl) {
-      if (!this.levelConfig) {
-        this.waveEl.textContent = '—';
-      } else {
-        if (this.isEndlessMode) {
-          const displayWave = this.combatActive
-            ? this.currentWaveNumber
-            : Math.max(1, this.currentWaveNumber || 1);
-          this.waveEl.textContent = `Wave ${displayWave}`;
-        } else {
-          const total = this.levelConfig.waves.length;
-          const displayWave = this.combatActive
-            ? this.waveIndex + 1
-            : Math.min(this.waveIndex + 1, total);
-          this.waveEl.textContent = `${displayWave}/${total}`;
-        }
-      }
-    }
-
-    if (this.healthEl) {
-      if (!this.levelConfig) {
-        this.healthEl.textContent = '—';
-      } else {
-        const currentLives = formatCombatNumber(this.lives);
-        const totalLives = formatCombatNumber(this.levelConfig.lives);
-        this.healthEl.textContent = `${currentLives}/${totalLives}`;
-      }
-    }
-
-    if (this.energyEl) {
-      if (!this.levelConfig) {
-        this.energyEl.textContent = '—';
-      } else if (!Number.isFinite(this.energy)) {
-        this.energyEl.textContent = `∞ ${this.theroSymbol}`;
-      } else {
-        const energyLabel = formatCombatNumber(this.energy);
-        this.energyEl.textContent = `${energyLabel} ${this.theroSymbol}`;
-      }
-    }
-
-    this.updateSpeedButton();
-    this.updateAutoAnchorButton();
-    refreshTowerLoadoutDisplay();
-    this.dependencies.updateStatusDisplays();
+    return HudBindings.updateHud.call(this);
   }
 
   updateProgress() {
-    if (!this.progressEl) {
-      return;
-    }
-
-    if (!this.levelConfig) {
-      this.progressEl.textContent = 'No active level.';
-      return;
-    }
-
-    if (!this.combatActive) {
-      if (this.resolvedOutcome === 'victory') {
-        const title = this.levelConfig.displayName || 'Defense';
-        this.progressEl.textContent = `${title} stabilized—victory sealed.`;
-      } else if (this.resolvedOutcome === 'defeat') {
-        const waveNote = this.maxWaveReached > 0 ? ` Reached wave ${this.maxWaveReached}.` : '';
-        this.progressEl.textContent = `Defense collapsed—rebuild the proof lattice.${waveNote}`;
-      } else {
-        const remainingMs =
-          this.autoWaveEnabled && this.autoStartDeadline
-            ? this.autoStartDeadline - Date.now()
-            : 0;
-        if (remainingMs > 0) {
-          const seconds = Math.max(0, Math.ceil(remainingMs / 1000));
-          const intro = this.isEndlessMode ? 'Endless mode primed' : 'Wave prep underway';
-          this.progressEl.textContent = `${intro}—auto-start in ${seconds}s.`;
-        } else {
-          this.progressEl.textContent = this.isEndlessMode
-            ? 'Endless mode primed—auto-start will trigger after preparations.'
-            : 'Wave prep underway.';
-        }
-      }
-      return;
-    }
-
-    const total = this.levelConfig.waves.length;
-    const remainingInWave = this.activeWave
-      ? Math.max(0, this.activeWave.config.count - this.activeWave.spawned)
-      : 0;
-    const remaining = remainingInWave + this.enemies.length;
-    const label = this.levelConfig.waves[this.waveIndex]?.label || 'glyphs';
-    if (this.isEndlessMode) {
-      this.progressEl.textContent = `Wave ${this.currentWaveNumber} — ${remaining} ${label} remaining.`;
-    } else {
-      const current = Math.min(this.waveIndex + 1, total);
-      this.progressEl.textContent = `Wave ${current}/${total} — ${remaining} ${label} remaining.`;
-    }
+    return HudBindings.updateProgress.call(this);
   }
 
   getCanvasPosition(normalized) {
@@ -6391,395 +5384,37 @@ export class SimplePlayfield {
   }
 
   draw() {
-    if (!this.ctx) {
-      return;
-    }
-    const ctx = this.ctx;
-    ctx.setTransform(1, 0, 0, 1, 0, 0);
-    ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
-    ctx.setTransform(this.pixelRatio, 0, 0, this.pixelRatio, 0, 0);
-
-    const width = this.renderWidth || (this.canvas ? this.canvas.clientWidth : 0) || 0;
-    const height = this.renderHeight || (this.canvas ? this.canvas.clientHeight : 0) || 0;
-    const viewCenter = this.getViewCenter();
-    ctx.translate(width / 2, height / 2);
-    ctx.scale(this.viewScale, this.viewScale);
-    ctx.translate(-viewCenter.x, -viewCenter.y);
-
-    this.drawFloaters();
-    this.drawPath();
-    this.drawMoteGems();
-    this.drawArcLight();
-    this.drawDeveloperCrystals();
-    this.drawNodes();
-    this.drawDeveloperPathMarkers();
-    this.drawPlacementPreview();
-    this.drawTowers();
-    this.drawDeltaSoldiers();
-    this.drawEnemies();
-    this.drawProjectiles();
-    this.drawTowerMenu();
-    this.updateEnemyTooltipPosition();
+    return CanvasRenderer.draw.call(this);
   }
 
   drawFloaters() {
-    if (!this.ctx || !this.floaters.length || !this.levelConfig) {
-      return;
-    }
-    const width = this.renderWidth || (this.canvas ? this.canvas.clientWidth : 0) || 0;
-    const height = this.renderHeight || (this.canvas ? this.canvas.clientHeight : 0) || 0;
-    if (!width || !height) {
-      return;
-    }
-    const minDimension = Math.min(width, height) || 1;
-    const connectionWidth = Math.max(0.6, minDimension * 0.0014);
-
-    const ctx = this.ctx;
-    ctx.save();
-    ctx.lineCap = 'round';
-    ctx.lineJoin = 'round';
-
-    this.floaterConnections.forEach((connection) => {
-      const from = this.floaters[connection.from];
-      const to = this.floaters[connection.to];
-      if (!from || !to) {
-        return;
-      }
-      const alpha = Math.max(0, Math.min(1, connection.strength || 0)) * 0.25;
-      if (alpha <= 0) {
-        return;
-      }
-      ctx.strokeStyle = `rgba(255, 255, 255, ${alpha})`;
-      ctx.lineWidth = connectionWidth;
-      ctx.beginPath();
-      ctx.moveTo(from.x, from.y);
-      ctx.lineTo(to.x, to.y);
-      ctx.stroke();
-    });
-
-    this.floaters.forEach((floater) => {
-      const opacity = Math.max(0, Math.min(1, floater.opacity || 0));
-      if (opacity <= 0) {
-        return;
-      }
-      let radiusFactor = Number.isFinite(floater.radiusFactor)
-        ? floater.radiusFactor
-        : null;
-      if (!radiusFactor) {
-        radiusFactor = this.randomFloaterRadiusFactor();
-        floater.radiusFactor = radiusFactor;
-      }
-      const radius = Math.max(2, radiusFactor * minDimension);
-      const strokeWidth = Math.max(0.8, radius * 0.22);
-      ctx.beginPath();
-      ctx.lineWidth = strokeWidth;
-      ctx.strokeStyle = `rgba(255, 255, 255, ${opacity * 0.25})`;
-      ctx.arc(floater.x, floater.y, radius, 0, Math.PI * 2);
-      ctx.stroke();
-    });
-
-    ctx.restore();
+    return CanvasRenderer.drawFloaters.call(this);
   }
 
   // Render each mote gem using its sprite when available so drops mirror the inventory art.
   drawMoteGems() {
-    if (!this.ctx || !moteGemState.active.length) {
-      return;
-    }
-    const ctx = this.ctx;
-    ctx.save();
-    const width = this.renderWidth || (this.canvas ? this.canvas.clientWidth : 0) || 0;
-    const height = this.renderHeight || (this.canvas ? this.canvas.clientHeight : 0) || 0;
-    const dimensionCandidates = [];
-    if (Number.isFinite(width) && width > 0) {
-      dimensionCandidates.push(width);
-    }
-    if (Number.isFinite(height) && height > 0) {
-      dimensionCandidates.push(height);
-    }
-    const minDimension = Math.max(
-      1,
-      dimensionCandidates.length ? Math.min(...dimensionCandidates) : 320,
-    );
-    const moteUnit = Math.max(6, minDimension * GEM_MOTE_BASE_RATIO);
-    const pulseMagnitude = moteUnit * 0.35;
-
-    moteGemState.active.forEach((gem) => {
-      const hue = gem.color?.hue ?? 48;
-      const saturation = gem.color?.saturation ?? 68;
-      const lightness = gem.color?.lightness ?? 56;
-      const moteSize = Number.isFinite(gem.moteSize) ? Math.max(1, gem.moteSize) : Math.max(1, gem.value);
-      const size = moteSize * moteUnit;
-      const pulse = Math.sin((gem.pulse || 0) * 0.6) * pulseMagnitude;
-      const rotation = Math.sin((gem.pulse || 0) * 0.35) * 0.45;
-      const opacity = Number.isFinite(gem.opacity) ? Math.max(0, Math.min(1, gem.opacity)) : 1;
-      const alphaFill = Math.max(0, Math.min(0.9, 0.6 + opacity * 0.3));
-      const alphaStroke = Math.max(0, Math.min(0.9, 0.5 + opacity * 0.35));
-      const fill = `hsla(${hue}, ${saturation}%, ${lightness}%, ${alphaFill})`;
-      const stroke = `hsla(${hue}, ${Math.max(24, saturation - 18)}%, ${Math.max(18, lightness - 28)}%, ${alphaStroke})`;
-      const sparkle = `hsla(${hue}, ${Math.max(34, saturation - 22)}%, 92%, ${Math.max(0, opacity * 0.65)})`;
-      const sprite = getGemSpriteImage(gem.typeKey);
-
-      ctx.save();
-      ctx.translate(gem.x, gem.y);
-      ctx.rotate(rotation);
-      if (sprite) {
-        // Scale the sprite according to mote size so drops stay legible on the battlefield.
-        const baseSize = Math.max(moteUnit * 0.6, size + pulse);
-        const reference = Math.max(1, Math.max(sprite.width || 1, sprite.height || 1));
-        const renderSize = baseSize;
-        const scale = renderSize / reference;
-        const width = (sprite.width || reference) * scale;
-        const height = (sprite.height || reference) * scale;
-        ctx.globalAlpha = opacity;
-        ctx.drawImage(sprite, -width / 2, -height / 2, width, height);
-      } else {
-        // Fall back to the square rendering when the sprite has not finished loading yet.
-        const squareSize = Math.max(moteUnit * 0.6, size + pulse);
-        const half = squareSize / 2;
-        ctx.fillStyle = fill;
-        ctx.strokeStyle = stroke;
-        ctx.lineWidth = Math.max(moteUnit * 0.12, 1.2);
-        ctx.beginPath();
-        ctx.rect(-half, -half, squareSize, squareSize);
-        ctx.fill();
-        ctx.stroke();
-
-        const sparkleSize = Math.max(moteUnit * 0.3, squareSize * 0.38);
-        ctx.fillStyle = sparkle;
-        ctx.fillRect(-sparkleSize * 0.5, -sparkleSize * 0.8, sparkleSize, sparkleSize);
-      }
-      ctx.restore();
-    });
-    ctx.restore();
+    return CanvasRenderer.drawMoteGems.call(this);
   }
 
   drawPath() {
-    if (!this.ctx || !this.pathSegments.length || this.pathPoints.length < 2) {
-      return;
-    }
-    const ctx = this.ctx;
-    const points = this.pathPoints;
-    const start = points[0];
-    const end = points[points.length - 1];
-
-    const baseGradient = ctx.createLinearGradient(start.x, start.y, end.x, end.y);
-    baseGradient.addColorStop(0, 'rgba(88, 160, 255, 0.5)');
-    baseGradient.addColorStop(0.48, 'rgba(162, 110, 255, 0.48)');
-    baseGradient.addColorStop(1, 'rgba(255, 158, 88, 0.5)');
-
-    ctx.save();
-    ctx.beginPath();
-    ctx.lineCap = 'round';
-    ctx.lineJoin = 'round';
-    ctx.lineWidth = 7;
-    this.applyCanvasShadow(ctx, 'rgba(88, 160, 255, 0.2)', 12);
-    ctx.moveTo(start.x, start.y);
-    for (let index = 1; index < points.length; index += 1) {
-      const point = points[index];
-      ctx.lineTo(point.x, point.y);
-    }
-    ctx.strokeStyle = baseGradient;
-    ctx.stroke();
-    ctx.restore();
-
-    const highlightGradient = ctx.createLinearGradient(start.x, start.y, end.x, end.y);
-    highlightGradient.addColorStop(0, 'rgba(88, 160, 255, 0.12)');
-    highlightGradient.addColorStop(0.52, 'rgba(162, 110, 255, 0.1)');
-    highlightGradient.addColorStop(1, 'rgba(255, 158, 88, 0.14)');
-
-    ctx.save();
-    ctx.beginPath();
-    ctx.lineCap = 'round';
-    ctx.lineJoin = 'round';
-    ctx.lineWidth = 2;
-    ctx.moveTo(start.x, start.y);
-    for (let index = 1; index < points.length; index += 1) {
-      const point = points[index];
-      ctx.lineTo(point.x, point.y);
-    }
-    ctx.strokeStyle = highlightGradient;
-    ctx.stroke();
-    ctx.restore();
+    return CanvasRenderer.drawPath.call(this);
   }
 
   drawArcLight() {
-    if (!this.ctx || !this.pathSegments.length || this.pathPoints.length < 2) {
-      return;
-    }
-    const ctx = this.ctx;
-    ctx.save();
-    ctx.beginPath();
-    ctx.lineCap = 'round';
-    ctx.lineJoin = 'round';
-    ctx.lineWidth = 3;
-    ctx.strokeStyle = 'rgba(255, 180, 105, 0.7)';
-    ctx.setLineDash([this.pathLength * 0.12, this.pathLength * 0.18]);
-    ctx.lineDashOffset = this.arcOffset;
-    ctx.moveTo(this.pathPoints[0].x, this.pathPoints[0].y);
-    for (let index = 1; index < this.pathPoints.length; index += 1) {
-      const point = this.pathPoints[index];
-      ctx.lineTo(point.x, point.y);
-    }
-    ctx.stroke();
-    ctx.restore();
+    return CanvasRenderer.drawArcLight.call(this);
   }
 
   // Render the enemy spawn gate using the dedicated sprite with a fallback glyph ring.
   drawEnemyGateSymbol(ctx, position) {
-    if (!ctx || !position) {
-      return;
-    }
-
-    const dimension = Math.min(this.renderWidth || 0, this.renderHeight || 0) || 0;
-    const baseRadius = dimension ? dimension * 0.028 : 0;
-    const baseSize = Math.max(12, Math.min(20, baseRadius || 16));
-    const radius = baseSize * 2;
-
-    ctx.save();
-    ctx.translate(position.x, position.y);
-
-    const glow = ctx.createRadialGradient(0, 0, radius * 0.2, 0, 0, radius * 1.2);
-    glow.addColorStop(0, 'rgba(74, 240, 255, 0.42)');
-    glow.addColorStop(1, 'rgba(15, 27, 63, 0)');
-    ctx.fillStyle = glow;
-    ctx.beginPath();
-    ctx.arc(0, 0, radius * 1.1, 0, Math.PI * 2);
-    ctx.fill();
-
-    const spriteReady = enemyGateSprite?.complete && enemyGateSprite.naturalWidth > 0;
-    if (spriteReady) {
-      // Anchor the enemy gate sprite at the path origin to match the Endless codex art.
-      const spriteSize = Math.max(baseSize * 2, 40) * 2;
-      ctx.save();
-      ctx.globalAlpha = 0.95;
-      ctx.drawImage(enemyGateSprite, -spriteSize / 2, -spriteSize / 2, spriteSize, spriteSize);
-      ctx.restore();
-    } else {
-      // Fallback portal composed of mirrored bezier arcs while the sprite loads.
-      this.applyCanvasShadow(ctx, 'rgba(74, 240, 255, 0.6)', radius * 0.6);
-      ctx.strokeStyle = 'rgba(202, 245, 255, 0.8)';
-      ctx.lineWidth = Math.max(1.6, radius * 0.14);
-      ctx.beginPath();
-      ctx.moveTo(-radius * 0.72, -radius * 0.1);
-      ctx.quadraticCurveTo(0, -radius * 0.8, radius * 0.72, -radius * 0.1);
-      ctx.quadraticCurveTo(0, radius * 0.6, -radius * 0.72, -radius * 0.1);
-      ctx.stroke();
-    }
-
-    ctx.restore();
+    return CanvasRenderer.drawEnemyGateSymbol.call(this, ctx, position);
   }
 
   drawMindGateSymbol(ctx, position) {
-    if (!ctx || !position) {
-      return;
-    }
-
-    const dimension = Math.min(this.renderWidth || 0, this.renderHeight || 0) || 0;
-    const baseRadius = dimension ? dimension * 0.035 : 0;
-    const baseSize = Math.max(14, Math.min(24, baseRadius || 18));
-    const radius = baseSize * 2;
-
-    ctx.save();
-    ctx.translate(position.x, position.y);
-
-    const glow = ctx.createRadialGradient(0, 0, radius * 0.22, 0, 0, radius);
-    glow.addColorStop(0, 'rgba(255, 248, 220, 0.9)');
-    glow.addColorStop(0.55, 'rgba(255, 196, 150, 0.35)');
-    glow.addColorStop(1, 'rgba(139, 247, 255, 0.18)');
-    ctx.fillStyle = glow;
-    ctx.beginPath();
-    ctx.arc(0, 0, radius, 0, Math.PI * 2);
-    ctx.fill();
-
-    this.applyCanvasShadow(ctx, 'rgba(255, 228, 120, 0.55)', radius);
-    ctx.strokeStyle = 'rgba(255, 228, 120, 0.85)';
-    ctx.lineWidth = Math.max(2, radius * 0.12);
-    ctx.beginPath();
-    ctx.arc(0, 0, radius * 0.88, 0, Math.PI * 2);
-    ctx.stroke();
-
-    const spriteReady = mindGateSprite?.complete && mindGateSprite.naturalWidth > 0;
-    if (spriteReady) {
-      // Draw the Mind Gate rune sprite scaled to the battlefield finale.
-      const spriteSize = Math.max(baseSize * 2.1, 46) * 2;
-      ctx.save();
-      ctx.globalAlpha = 0.96;
-      ctx.drawImage(mindGateSprite, -spriteSize / 2, -spriteSize / 2, spriteSize, spriteSize);
-      ctx.restore();
-
-      // Removed inner blue ring to avoid small light-blue circle at the track end.
-    } else {
-      // Fall back to the legacy geometric rendering until the sprite finishes loading.
-      this.applyCanvasShadow(ctx, 'rgba(139, 247, 255, 0.55)', radius * 0.7);
-      ctx.strokeStyle = 'rgba(139, 247, 255, 0.85)';
-      ctx.lineWidth = Math.max(1.4, radius * 0.12);
-      ctx.beginPath();
-      ctx.moveTo(0, radius * 0.64);
-      ctx.lineTo(0, -radius * 0.6);
-      ctx.stroke();
-
-      // Removed inner blue ring to avoid small light-blue circle at the track end.
-
-      ctx.strokeStyle = 'rgba(255, 228, 120, 0.92)';
-      this.applyCanvasShadow(ctx, 'rgba(255, 228, 120, 0.55)', radius * 0.8);
-      ctx.lineWidth = Math.max(1.6, radius * 0.14);
-      ctx.beginPath();
-      const gateWidth = radius * 0.58;
-      const gateBase = radius * 0.62;
-      ctx.moveTo(-gateWidth, gateBase);
-      ctx.lineTo(-gateWidth, -radius * 0.18);
-      ctx.quadraticCurveTo(0, -radius * 0.95, gateWidth, -radius * 0.18);
-      ctx.lineTo(gateWidth, gateBase);
-      ctx.stroke();
-    }
-
-    const gateIntegrity = Math.max(0, Math.floor(this.lives || 0));
-    const maxIntegrity = Math.max(
-      gateIntegrity,
-      Math.floor(this.levelConfig?.lives || gateIntegrity || 1),
-    );
-    const gateExponentSource = gateIntegrity > 0 ? gateIntegrity : maxIntegrity || 1;
-    const gateExponent = this.calculateHealthExponent(gateExponentSource);
-    const palette =
-      typeof this.getEffectiveMotePalette === 'function'
-        ? this.getEffectiveMotePalette()
-        : null;
-    const paletteStops = resolvePaletteColorStops(palette);
-    const gradient = ctx.createLinearGradient(-radius, -radius, radius, radius);
-    if (Array.isArray(paletteStops) && paletteStops.length) {
-      const denominator = Math.max(1, paletteStops.length - 1);
-      paletteStops.forEach((stop, index) => {
-        const offset = Math.max(0, Math.min(1, index / denominator));
-        gradient.addColorStop(offset, colorToRgbaString(stop, 1));
-      });
-    }
-    ctx.font = `${Math.round(Math.max(14, radius * 0.82))}px "Cormorant Garamond", serif`;
-    ctx.textAlign = 'left';
-    ctx.textBaseline = 'bottom';
-    ctx.fillStyle = gradient;
-    const highlightColor = paletteStops[paletteStops.length - 1] || paletteStops[0];
-    this.applyCanvasShadow(ctx, colorToRgbaString(highlightColor, 0.85), Math.max(14, radius * 0.95));
-    const exponentOffset = radius * 0.78;
-    const exponentX = exponentOffset;
-    const exponentY = -exponentOffset * 0.88;
-    // Show the core's exponent with the same decimal precision used for enemy glyphs.
-    ctx.fillText(gateExponent.toFixed(1), exponentX, exponentY);
-
-    ctx.restore();
+    return CanvasRenderer.drawMindGateSymbol.call(this, ctx, position);
   }
 
   drawNodes() {
-    if (!this.ctx || !this.pathSegments.length) {
-      return;
-    }
-    const ctx = this.ctx;
-    const startPoint = this.pathPoints.length ? this.pathPoints[0] : this.pathSegments[0].start;
-    const endPoint = this.pathPoints.length
-      ? this.pathPoints[this.pathPoints.length - 1]
-      : this.pathSegments[this.pathSegments.length - 1].end;
-    this.drawEnemyGateSymbol(ctx, startPoint);
-    this.drawMindGateSymbol(ctx, endPoint);
+    return CanvasRenderer.drawNodes.call(this);
   }
 
   setDeveloperPathMarkers(markers) {
@@ -6812,43 +5447,7 @@ export class SimplePlayfield {
   }
 
   drawDeveloperPathMarkers() {
-    if (!this.ctx || !Array.isArray(this.developerPathMarkers) || !this.developerPathMarkers.length) {
-      return;
-    }
-
-    const ctx = this.ctx;
-    ctx.save();
-    ctx.font = '12px "Cormorant Garamond", serif';
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
-
-    this.developerPathMarkers.forEach((marker, index) => {
-      const radius = marker.active ? 12 : 10;
-      ctx.beginPath();
-      ctx.fillStyle = marker.active ? 'rgba(18, 26, 44, 0.9)' : 'rgba(12, 16, 28, 0.82)';
-      ctx.strokeStyle = marker.active
-        ? 'rgba(139, 247, 255, 0.9)'
-        : 'rgba(139, 247, 255, 0.55)';
-      ctx.lineWidth = marker.active ? 2 : 1.5;
-      if (marker.active) {
-        this.applyCanvasShadow(ctx, 'rgba(139, 247, 255, 0.3)', 16);
-      } else {
-        this.clearCanvasShadow(ctx);
-      }
-      ctx.arc(marker.x, marker.y, radius, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.stroke();
-
-      const label = marker.label !== undefined && marker.label !== null ? marker.label : index + 1;
-      if (label !== undefined && label !== null) {
-        ctx.fillStyle = 'rgba(139, 247, 255, 0.9)';
-        ctx.fillText(String(label), marker.x, marker.y);
-      }
-    });
-
-
-    // Restore the canvas state so developer marker styling does not leak into other drawing routines.
-    ctx.restore();
+    return CanvasRenderer.drawDeveloperPathMarkers.call(this);
   }
 
   // Resolve a developer crystal's radius relative to the active render dimensions.
@@ -7119,769 +5718,71 @@ export class SimplePlayfield {
 
   // Render developer crystals and their shard debris above the lane.
   drawDeveloperCrystals() {
-    if (!this.ctx) {
-      return;
-    }
-    const ctx = this.ctx;
-    if (this.developerCrystals.length) {
-      ctx.save();
-      this.developerCrystals.forEach((crystal) => {
-        if (!crystal) {
-          return;
-        }
-        const position = this.getCrystalPosition(crystal);
-        const radius = this.getCrystalRadius(crystal);
-        if (!position || radius <= 0) {
-          return;
-        }
-        ctx.save();
-        ctx.translate(position.x, position.y);
-        ctx.rotate(crystal.orientation || 0);
-        const outline = Array.isArray(crystal.outline) && crystal.outline.length
-          ? crystal.outline
-          : [1, 1, 1, 1, 1, 1];
-        ctx.beginPath();
-        outline.forEach((scale, index) => {
-          const ratio = index / outline.length;
-          const angle = ratio * Math.PI * 2;
-          const radial = radius * (0.72 + scale * 0.28);
-          const x = Math.cos(angle) * radial;
-          const y = Math.sin(angle) * radial;
-          if (index === 0) {
-            ctx.moveTo(x, y);
-          } else {
-            ctx.lineTo(x, y);
-          }
-        });
-        ctx.closePath();
-        const baseColor = samplePaletteGradient(crystal.paletteRatio ?? 0.5) || { r: 160, g: 220, b: 255 };
-        const highlightColor = samplePaletteGradient(Math.min(1, (crystal.paletteRatio ?? 0.5) + 0.18)) || baseColor;
-        const gradient = ctx.createLinearGradient(-radius, -radius, radius, radius);
-        gradient.addColorStop(0, colorToRgbaString(baseColor, 0.88));
-        gradient.addColorStop(1, colorToRgbaString(highlightColor, 0.82));
-        ctx.fillStyle = gradient;
-        ctx.fill();
-        ctx.lineWidth = Math.max(2, radius * 0.08);
-        ctx.strokeStyle = 'rgba(255, 255, 255, 0.12)';
-        ctx.stroke();
-        if (Array.isArray(crystal.fractures)) {
-          crystal.fractures.forEach((fracture) => {
-            if (!fracture) {
-              return;
-            }
-            const width = Number.isFinite(fracture.width) ? fracture.width : 0.6;
-            const depth = Number.isFinite(fracture.depth) ? fracture.depth : 0.4;
-            const progress = Number.isFinite(fracture.progress) ? fracture.progress : 0;
-            const jagged = Array.isArray(fracture.jagged) && fracture.jagged.length
-              ? fracture.jagged
-              : [1, 0.7, 0.85, 0.7, 1];
-            const segments = Math.max(3, jagged.length - 1);
-            const outerRadius = radius * 0.98;
-            ctx.save();
-            const anchor = Number.isFinite(fracture.angle) ? fracture.angle : 0;
-            ctx.rotate(anchor);
-            ctx.beginPath();
-            for (let index = 0; index <= segments; index += 1) {
-              const t = index / segments;
-              const angle = (t - 0.5) * width;
-              const x = Math.cos(angle) * outerRadius;
-              const y = Math.sin(angle) * outerRadius;
-              if (index === 0) {
-                ctx.moveTo(x, y);
-              } else {
-                ctx.lineTo(x, y);
-              }
-            }
-            for (let index = segments; index >= 0; index -= 1) {
-              const t = index / segments;
-              const angle = (t - 0.5) * width;
-              const jaggedScale = jagged[index] ?? 1;
-              const inset = Math.min(0.9, depth * progress * jaggedScale);
-              const radial = outerRadius * (1 - inset);
-              const x = Math.cos(angle) * radial;
-              const y = Math.sin(angle) * radial;
-              ctx.lineTo(x, y);
-            }
-            ctx.closePath();
-            ctx.fillStyle = 'rgba(8, 12, 24, 0.9)';
-            ctx.fill();
-            ctx.restore();
-          });
-        }
-        if (this.focusedCrystalId === crystal.id) {
-          ctx.strokeStyle = 'rgba(255, 228, 120, 0.85)';
-          ctx.setLineDash([6, 6]);
-          ctx.lineWidth = 2;
-          ctx.beginPath();
-          ctx.arc(0, 0, radius * 1.12, 0, Math.PI * 2);
-          ctx.stroke();
-          ctx.setLineDash([]);
-        }
-        ctx.restore();
-      });
-      ctx.restore();
-    }
-    if (this.crystalShards.length) {
-      ctx.save();
-      this.crystalShards.forEach((shard) => {
-        if (!shard) {
-          return;
-        }
-        const lifeRatio = shard.maxLife ? Math.max(0, 1 - shard.life / shard.maxLife) : 1;
-        const shardColor = colorToRgbaString(shard.color || { r: 188, g: 236, b: 255 }, lifeRatio);
-        ctx.save();
-        ctx.translate(shard.x || 0, shard.y || 0);
-        ctx.rotate(shard.rotation || 0);
-        ctx.fillStyle = shardColor;
-        const size = Math.max(2, shard.size || 5);
-        ctx.beginPath();
-        ctx.moveTo(0, -size * 0.4);
-        ctx.lineTo(size * 0.6, 0);
-        ctx.lineTo(0, size * 0.5);
-        ctx.closePath();
-        ctx.fill();
-        ctx.restore();
-      });
-      ctx.restore();
-    }
+    return CanvasRenderer.drawDeveloperCrystals.call(this);
   }
 
   drawPlacementPreview() {
-    if (!this.ctx || !this.hoverPlacement || !this.hoverPlacement.position) {
-      return;
-    }
-
-    const ctx = this.ctx;
-    const {
-      position,
-      range,
-      valid,
-      merge,
-      mergeTarget,
-      symbol,
-      reason,
-      dragging,
-      towerType,
-      definition,
-      tier,
-    } = this.hoverPlacement;
-
-    ctx.save();
-
-    const radius = Number.isFinite(range) && range > 0 ? range : Math.min(this.renderWidth, this.renderHeight) * 0.18;
-    const fillColor = valid ? 'rgba(139, 247, 255, 0.12)' : 'rgba(255, 112, 112, 0.16)';
-    const strokeColor = valid ? 'rgba(139, 247, 255, 0.85)' : 'rgba(255, 96, 96, 0.9)';
-
-    // Render the projected range circle so players can gauge coverage during placement.
-    ctx.beginPath();
-    ctx.fillStyle = fillColor;
-    ctx.strokeStyle = strokeColor;
-    ctx.lineWidth = valid ? 2 : 3;
-    ctx.arc(position.x, position.y, Math.max(12, radius), 0, Math.PI * 2);
-    ctx.fill();
-    ctx.stroke();
-
-    if (merge && mergeTarget) {
-      // Highlight the merge target to reinforce that the placement will combine towers.
-      ctx.setLineDash([6, 6]);
-      ctx.lineWidth = 2;
-      ctx.strokeStyle = 'rgba(255, 236, 128, 0.85)';
-      ctx.beginPath();
-      ctx.arc(mergeTarget.x, mergeTarget.y, Math.max(16, (radius || 24) * 0.6), 0, Math.PI * 2);
-      ctx.stroke();
-      ctx.setLineDash([]);
-    }
-
-    const previewDefinition = definition || getTowerDefinition(towerType);
-    const previewTower = {
-      type: towerType,
-      definition: previewDefinition || undefined,
-      tier: Number.isFinite(tier) ? tier : previewDefinition?.tier,
-      symbol,
-    };
-    const visuals = getTowerVisualConfig(previewTower) || {};
-    const bodyRadius = Math.max(
-      12,
-      Math.min(this.renderWidth, this.renderHeight) * ALPHA_BASE_RADIUS_FACTOR,
-    );
-    const bodyStroke = valid
-      ? visuals.outerStroke || 'rgba(139, 247, 255, 0.85)'
-      : 'rgba(255, 96, 96, 0.85)';
-    const bodyFill = valid
-      ? visuals.innerFill || 'rgba(12, 16, 28, 0.9)'
-      : 'rgba(60, 16, 16, 0.88)';
-    const symbolFill = valid
-      ? visuals.symbolFill || 'rgba(255, 228, 120, 0.85)'
-      : 'rgba(255, 200, 200, 0.92)';
-
-    ctx.save();
-    if (valid && visuals.outerShadow?.color) {
-      this.applyCanvasShadow(
-        ctx,
-        visuals.outerShadow.color,
-        Number.isFinite(visuals.outerShadow.blur) ? visuals.outerShadow.blur : 18,
-      );
-    } else {
-      this.clearCanvasShadow(ctx);
-    }
-    ctx.beginPath();
-    ctx.fillStyle = bodyFill;
-    ctx.strokeStyle = bodyStroke;
-    ctx.lineWidth = valid ? 2.4 : 2.6;
-    ctx.arc(position.x, position.y, bodyRadius, 0, Math.PI * 2);
-    ctx.fill();
-    ctx.stroke();
-    ctx.restore();
-
-    ctx.save();
-    if (valid && visuals.symbolShadow?.color) {
-      this.applyCanvasShadow(
-        ctx,
-        visuals.symbolShadow.color,
-        Number.isFinite(visuals.symbolShadow.blur) ? visuals.symbolShadow.blur : 18,
-      );
-    } else {
-      this.clearCanvasShadow(ctx);
-    }
-    const glyph = symbol || '·';
-    ctx.font = `${Math.round(bodyRadius * 1.4)}px "Cormorant Garamond", serif`;
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
-    ctx.fillStyle = symbolFill;
-    ctx.fillText(glyph, position.x, position.y);
-    ctx.restore();
-
-    if (dragging) {
-      // Add a subtle outline while dragging to indicate the pointer anchor.
-      ctx.lineWidth = 1.5;
-      ctx.strokeStyle = 'rgba(139, 247, 255, 0.4)';
-      ctx.beginPath();
-      const anchorRadius = Math.max(bodyRadius * 1.15, bodyRadius + 4, 16);
-      ctx.arc(position.x, position.y, anchorRadius, 0, Math.PI * 2);
-      ctx.stroke();
-    }
-
-    ctx.restore();
-
-    if (this.messageEl && reason) {
-      // Surface placement feedback in the UI message element for accessibility.
-      this.messageEl.textContent = reason;
-    }
+    return CanvasRenderer.drawPlacementPreview.call(this);
   }
 
   drawTowers() {
-    if (!this.ctx || !this.towers.length) {
-      return;
-    }
-
-    const ctx = this.ctx;
-    ctx.save();
-
-    this.drawConnectionEffects(ctx);
-
-    const activeDrag = this.connectionDragState.active ? this.connectionDragState : null;
-    const highlightEntries = activeDrag && Array.isArray(activeDrag.highlightEntries)
-      ? activeDrag.highlightEntries
-      : [];
-    const highlightMap = new Map();
-    highlightEntries.forEach((entry) => {
-      if (!highlightMap.has(entry.towerId)) {
-        highlightMap.set(entry.towerId, entry);
-      }
-    });
-    const hoveredHighlight = activeDrag ? activeDrag.hoverEntry : null;
-
-    this.towers.forEach((tower) => {
-      if (!tower || !Number.isFinite(tower.x) || !Number.isFinite(tower.y)) {
-        return;
-      }
-
-      const visuals = getTowerVisualConfig(tower) || {};
-      const rangeRadius = Number.isFinite(tower.range)
-        ? tower.range
-        : Math.min(this.renderWidth, this.renderHeight) * 0.22;
-      const bodyRadius = Math.max(
-        12,
-        Math.min(this.renderWidth, this.renderHeight) * ALPHA_BASE_RADIUS_FACTOR,
-      );
-
-      const highlightEntry = highlightMap.get(tower.id) || null;
-      if (highlightEntry) {
-        ctx.save();
-        const isHovered = hoveredHighlight && hoveredHighlight.towerId === tower.id;
-        const strokeColor = highlightEntry.action === 'connect'
-          ? isHovered
-            ? 'rgba(139, 247, 255, 0.85)'
-            : 'rgba(139, 247, 255, 0.45)'
-          : isHovered
-          ? 'rgba(255, 214, 112, 0.85)'
-          : 'rgba(255, 214, 112, 0.45)';
-        ctx.lineWidth = isHovered ? 3.2 : 2;
-        ctx.strokeStyle = strokeColor;
-        ctx.setLineDash([isHovered ? 6 : 4, isHovered ? 6 : 8]);
-        ctx.beginPath();
-        ctx.arc(tower.x, tower.y, bodyRadius + 10, 0, Math.PI * 2);
-        ctx.stroke();
-        ctx.setLineDash([]);
-        ctx.restore();
-      }
-
-      // Outline the active coverage radius so players can gauge lattice reach in real time.
-      if (Number.isFinite(rangeRadius) && rangeRadius > 0) {
-        ctx.beginPath();
-        ctx.lineWidth = 1.5;
-        ctx.strokeStyle = visuals.rangeStroke || 'rgba(139, 247, 255, 0.2)';
-        ctx.setLineDash([8, 6]);
-        ctx.arc(tower.x, tower.y, rangeRadius, 0, Math.PI * 2);
-        ctx.stroke();
-        ctx.setLineDash([]);
-      }
-
-      if (tower.type === 'zeta') {
-        this.drawZetaPendulums(tower);
-      }
-      if (tower.type === 'eta') {
-        this.drawEtaOrbits(tower);
-      }
-
-      ctx.save();
-      const outerShadow = visuals.outerShadow;
-      if (outerShadow?.color) {
-        this.applyCanvasShadow(
-          ctx,
-          outerShadow.color,
-          Number.isFinite(outerShadow.blur) ? outerShadow.blur : 18,
-        );
-      } else {
-        this.clearCanvasShadow(ctx);
-      }
-
-      ctx.beginPath();
-      ctx.fillStyle = visuals.innerFill || 'rgba(12, 16, 28, 0.9)';
-      ctx.strokeStyle = visuals.outerStroke || 'rgba(139, 247, 255, 0.75)';
-      ctx.lineWidth = 2.4;
-      ctx.arc(tower.x, tower.y, bodyRadius, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.stroke();
-      ctx.restore();
-
-      this.drawTowerConnectionParticles(ctx, tower, bodyRadius);
-
-      const symbolColor = visuals.symbolFill || 'rgba(255, 228, 120, 0.92)';
-      const symbolShadow = visuals.symbolShadow;
-      if (symbolShadow?.color) {
-        this.applyCanvasShadow(
-          ctx,
-          symbolShadow.color,
-          Number.isFinite(symbolShadow.blur) ? symbolShadow.blur : 18,
-        );
-      } else {
-        this.clearCanvasShadow(ctx);
-      }
-
-      const glyph = tower.symbol || tower.definition?.symbol || '?';
-      ctx.fillStyle = symbolColor;
-      ctx.font = `${Math.round(bodyRadius * 1.4)}px "Cormorant Garamond", serif`;
-      ctx.textAlign = 'center';
-      ctx.textBaseline = 'middle';
-      ctx.fillText(glyph, tower.x, tower.y);
-
-      if (tower.type === 'beta') {
-        const alphaShots = Math.max(0, Math.floor(tower.storedAlphaShots || 0));
-        if (alphaShots > 0) {
-          ctx.save();
-          ctx.fillStyle = 'rgba(255, 255, 255, 0.9)';
-          ctx.font = `${Math.round(bodyRadius * 0.75)}px "Cormorant Garamond", serif`;
-          ctx.textAlign = 'center';
-          ctx.textBaseline = 'top';
-          ctx.fillText(`α × ${alphaShots}`, tower.x, tower.y + bodyRadius + 6);
-          ctx.restore();
-        }
-      } else if (tower.type === 'gamma') {
-        const betaShots = Math.max(0, Math.floor(tower.storedBetaShots || 0));
-        const alphaShots = Math.max(0, Math.floor(tower.storedAlphaShots || 0));
-        if (betaShots > 0 || alphaShots > 0) {
-          ctx.save();
-          ctx.fillStyle = 'rgba(255, 255, 255, 0.9)';
-          ctx.font = `${Math.round(bodyRadius * 0.7)}px "Cormorant Garamond", serif`;
-          ctx.textAlign = 'center';
-          ctx.textBaseline = 'top';
-          let labelY = tower.y + bodyRadius + 6;
-          if (betaShots > 0) {
-            ctx.fillText(`β × ${betaShots}`, tower.x, labelY);
-            labelY += Math.round(bodyRadius * 0.7) + 4;
-          }
-          if (alphaShots > 0) {
-            ctx.fillText(`α × ${alphaShots}`, tower.x, labelY);
-          }
-          ctx.restore();
-        }
-      }
-
-      if (tower.chain) {
-        // Highlight Aleph-linked towers with a secondary ring that pulses with the chain state.
-        this.applyCanvasShadow(ctx, 'rgba(255, 228, 120, 0.55)', 20);
-        ctx.strokeStyle = 'rgba(255, 228, 120, 0.75)';
-        ctx.lineWidth = 2;
-        ctx.beginPath();
-        ctx.arc(tower.x, tower.y, bodyRadius + 6, 0, Math.PI * 2);
-        ctx.stroke();
-        this.clearCanvasShadow(ctx);
-      }
-
-      if (this.activeTowerMenu?.towerId === tower.id) {
-        // Outline the selected lattice so the menu context is clear.
-        ctx.beginPath();
-        ctx.strokeStyle = 'rgba(139, 247, 255, 0.9)';
-        ctx.lineWidth = 2.6;
-        ctx.arc(tower.x, tower.y, bodyRadius + 10, 0, Math.PI * 2);
-        ctx.stroke();
-      }
-    });
-
-    ctx.restore();
+    return CanvasRenderer.drawTowers.call(this);
   }
 
   /**
    * Render ζ pendulum arms and trails so the battlefield reflects their orbit.
    */
   drawZetaPendulums(tower) {
-    drawZetaPendulumsHelper(this, tower);
+    return CanvasRenderer.drawZetaPendulums.call(this, tower);
   }
 
   /**
    * Render η orbital rings and trailing motes for the planetary lattice.
    */
   drawEtaOrbits(tower) {
-    drawEtaOrbitsHelper(this, tower);
+    return CanvasRenderer.drawEtaOrbits.call(this, tower);
   }
 
   /**
    * Render the roaming Delta soldiers as luminous triangles circling the battlefield.
    */
   drawDeltaSoldiers() {
-    drawDeltaSoldiersHelper(this);
+    return CanvasRenderer.drawDeltaSoldiers.call(this);
   }
 
   drawEnemies() {
-    if (!this.ctx || !this.enemies.length) {
-      return;
-    }
-
-    const ctx = this.ctx;
-    ctx.save();
-
-    this.enemies.forEach((enemy) => {
-      if (!enemy) {
-        return;
-      }
-
-      const position = this.getEnemyPosition(enemy);
-      if (!position) {
-        return;
-      }
-
-      const metrics = this.getEnemyVisualMetrics(enemy);
-      const symbol = typeof enemy.symbol === 'string' ? enemy.symbol : this.resolveEnemySymbol(enemy);
-      const exponent = this.calculateHealthExponent(Math.max(1, enemy.hp ?? enemy.maxHp ?? 1));
-
-      ctx.save();
-      ctx.translate(position.x, position.y);
-
-      ctx.beginPath();
-      ctx.fillStyle = 'rgba(12, 16, 24, 0.88)';
-      ctx.strokeStyle = 'rgba(139, 247, 255, 0.45)';
-      ctx.lineWidth = 2;
-      ctx.arc(0, 0, metrics.ringRadius, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.stroke();
-
-      ctx.beginPath();
-      ctx.fillStyle = 'rgba(139, 247, 255, 0.28)';
-      ctx.arc(0, 0, metrics.coreRadius, 0, Math.PI * 2);
-      ctx.fill();
-
-      ctx.fillStyle = 'rgba(255, 255, 255, 0.92)';
-      ctx.font = `${metrics.symbolSize}px "Cormorant Garamond", serif`;
-      ctx.textAlign = 'center';
-      ctx.textBaseline = 'middle';
-      ctx.fillText(symbol || '?', 0, 0);
-
-      ctx.font = `700 ${metrics.exponentSize}px "Cormorant Garamond", serif`;
-      ctx.textAlign = 'right';
-      ctx.textBaseline = 'top';
-      const exponentLabel = exponent.toFixed(1);
-      const exponentColor = this.resolveEnemyExponentColor(enemy);
-      // Match the outline width to a single CSS pixel so exponents stay crisp on high-DPI displays.
-      const pixelRatio = Number.isFinite(this.pixelRatio) && this.pixelRatio > 0 ? this.pixelRatio : 1;
-      const outlineWidth = Math.max(1, Math.round(pixelRatio));
-      ctx.lineJoin = 'round';
-      ctx.miterLimit = 2;
-      ctx.strokeStyle = 'rgba(6, 8, 14, 0.85)';
-      ctx.lineWidth = outlineWidth;
-      // Display the precise exponent so each glyph telegraphs its ten-power health tier.
-      const exponentOffsetX = metrics.ringRadius * 0.94;
-      const exponentOffsetY = -metrics.ringRadius * 0.98;
-      // Anchor the exponent to the enemy's top-right quadrant so the scientific notation reads cleanly above the glyph.
-      // Outline the exponent before filling so it remains legible over bright projectile bursts.
-      ctx.strokeText(exponentLabel, exponentOffsetX, exponentOffsetY);
-      // Paint the exponent with the threat-informed palette so players grasp incoming danger immediately.
-      ctx.fillStyle = exponentColor;
-      ctx.fillText(exponentLabel, exponentOffsetX, exponentOffsetY);
-
-      if (this.focusedEnemyId === enemy.id) {
-        // Draw a rotating focus marker so players can track the prioritized enemy.
-        const markerRadius = metrics.focusRadius || metrics.ringRadius + 8;
-        const angle = this.focusMarkerAngle || 0;
-        const span = Math.PI / 3;
-        ctx.strokeStyle = 'rgba(255, 228, 120, 0.85)';
-        ctx.lineWidth = 2;
-        ctx.beginPath();
-        ctx.arc(0, 0, markerRadius, angle, angle + span);
-        ctx.stroke();
-        ctx.beginPath();
-        ctx.arc(0, 0, markerRadius, angle + Math.PI, angle + Math.PI + span);
-        ctx.stroke();
-      }
-
-      ctx.restore();
-    });
-
-    ctx.restore();
+    return CanvasRenderer.drawEnemies.call(this);
   }
 
   drawProjectiles() {
-    if (!this.ctx) {
-      return;
-    }
-
-    const ctx = this.ctx;
-    if (this.projectiles.length) {
-      ctx.save();
-    }
-
-    this.projectiles.forEach((projectile) => {
-      if (!projectile) {
-        return;
-      }
-
-      if (projectile.patternType === 'supply') {
-        const position = projectile.currentPosition || projectile.target || projectile.source;
-        if (!position) {
-          return;
-        }
-        const seeds = Array.isArray(projectile.seeds) ? projectile.seeds : [];
-        if (seeds.length) {
-          seeds.forEach((seed) => {
-            if (!seed || !seed.position) {
-              return;
-            }
-            const baseColor = seed.type === 'beta'
-              ? { r: 255, g: 214, b: 112 }
-              : { r: 255, g: 138, b: 216 };
-            const glowColor = normalizeProjectileColor(baseColor, 1);
-            const size = Number.isFinite(seed.size) ? seed.size : 2.2;
-            ctx.fillStyle = colorToRgbaString(glowColor, 0.85);
-            ctx.beginPath();
-            ctx.arc(seed.position.x, seed.position.y, size, 0, Math.PI * 2);
-            ctx.fill();
-          });
-        } else {
-          const color = normalizeProjectileColor(projectile.color, 1);
-          ctx.fillStyle = colorToRgbaString(color, 0.85);
-          ctx.beginPath();
-          ctx.arc(position.x, position.y, 4, 0, Math.PI * 2);
-          ctx.fill();
-        }
-        return;
-      }
-
-      if (projectile.patternType === 'omegaWave') {
-        const position = projectile.position || projectile.origin;
-        if (!position) {
-          return;
-        }
-
-        const radius = Number.isFinite(projectile.parameters?.radius)
-          ? projectile.parameters.radius
-          : 40;
-        const gradient = ctx.createRadialGradient(position.x, position.y, 0, position.x, position.y, radius);
-        gradient.addColorStop(0, 'rgba(255, 228, 120, 0.8)');
-        gradient.addColorStop(1, 'rgba(255, 228, 120, 0)');
-
-        // Render swirling omega waves as glowing motes that fade across their path envelope.
-        ctx.fillStyle = gradient;
-        ctx.beginPath();
-        ctx.arc(position.x, position.y, radius, 0, Math.PI * 2);
-        ctx.fill();
-        return;
-      }
-
-      if (projectile.patternType === 'etaLaser') {
-        const origin = projectile.origin;
-        if (!origin) {
-          return;
-        }
-        const angle = Number.isFinite(projectile.angle) ? projectile.angle : 0;
-        const length = Number.isFinite(projectile.length) ? projectile.length : 0;
-        if (length <= 0) {
-          return;
-        }
-        const width = Math.max(2, Number.isFinite(projectile.width) ? projectile.width : 8);
-        const alpha = Number.isFinite(projectile.alpha) ? Math.max(0, Math.min(1, projectile.alpha)) : 1;
-        ctx.save();
-        ctx.translate(origin.x, origin.y);
-        ctx.rotate(angle);
-        const beamColor = normalizeProjectileColor(projectile.color, 1);
-        const gradient = ctx.createLinearGradient(0, 0, length, 0);
-        gradient.addColorStop(0, colorToRgbaString(beamColor, alpha));
-        gradient.addColorStop(0.6, colorToRgbaString(beamColor, alpha * 0.6));
-        gradient.addColorStop(1, colorToRgbaString(beamColor, 0));
-        ctx.strokeStyle = gradient;
-        ctx.lineWidth = width;
-        ctx.beginPath();
-        ctx.moveTo(0, 0);
-        ctx.lineTo(length, 0);
-        ctx.stroke();
-        ctx.restore();
-        return;
-      }
-
-      if (projectile.patternType === 'epsilonNeedle') {
-        const position = projectile.position || projectile.origin;
-        if (!position) {
-          return;
-        }
-        const prev = projectile.previousPosition || position;
-        const heading = Math.atan2((position.y - prev.y) || 0.0001, (position.x - prev.x) || 0.0001);
-        const length = 10;
-        const width = 2.2;
-        ctx.save();
-        ctx.translate(position.x, position.y);
-        ctx.rotate(heading);
-        ctx.fillStyle = 'rgba(139, 247, 255, 0.9)';
-        ctx.strokeStyle = 'rgba(12, 16, 26, 0.9)';
-        ctx.lineWidth = 1.2;
-        ctx.beginPath();
-        ctx.moveTo(length, 0);
-        ctx.lineTo(-length * 0.6, width);
-        ctx.lineTo(-length * 0.6, -width);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        ctx.restore();
-        return;
-      }
-
-      const source = projectile.source;
-      const targetPosition = projectile.target
-        ? projectile.target
-        : projectile.targetId
-        ? (() => {
-            const enemy = this.enemies.find((candidate) => candidate.id === projectile.targetId);
-            return enemy ? this.getEnemyPosition(enemy) : null;
-          })()
-        : projectile.targetCrystalId
-        ? (() => {
-            const crystal = this.developerCrystals.find((entry) => entry?.id === projectile.targetCrystalId);
-            return crystal ? this.getCrystalPosition(crystal) : null;
-          })()
-        : null;
-
-      if (!source || !targetPosition) {
-        return;
-      }
-
-      const beamStart = normalizeProjectileColor(projectile.color, 0);
-      const beamEnd = normalizeProjectileColor(projectile.color, 1);
-      const beamGradient = ctx.createLinearGradient(source.x, source.y, targetPosition.x, targetPosition.y);
-      beamGradient.addColorStop(0, colorToRgbaString(beamStart, 0.72));
-      beamGradient.addColorStop(1, colorToRgbaString(beamEnd, 0.78));
-      ctx.strokeStyle = beamGradient;
-      ctx.lineWidth = 1.5;
-      ctx.beginPath();
-      ctx.moveTo(source.x, source.y);
-      ctx.lineTo(targetPosition.x, targetPosition.y);
-      ctx.stroke();
-
-      ctx.fillStyle = colorToRgbaString(beamEnd, 0.9);
-      ctx.beginPath();
-      ctx.arc(targetPosition.x, targetPosition.y, 4, 0, Math.PI * 2);
-      ctx.fill();
-    });
-
-    if (this.projectiles.length) {
-      ctx.restore();
-    }
-
-    this.drawBetaBursts();
-    this.drawAlphaBursts();
-    this.drawGammaBursts();
+    return CanvasRenderer.drawProjectiles.call(this);
   }
 
   /**
    * Render α particle bursts as soft-glow motes swirling around the lattice.
    */
   drawAlphaBursts() {
-    drawAlphaBurstsHelper(this);
+    return CanvasRenderer.drawAlphaBursts.call(this);
   }
 
   /**
    * Render β particle bursts to visualize mirrored exponential energy.
    */
   drawBetaBursts() {
-    drawBetaBurstsHelper(this);
+    return CanvasRenderer.drawBetaBursts.call(this);
   }
 
   /**
    * Render γ particle bursts so piercing lasers remain visible in combat.
    */
   drawGammaBursts() {
-    drawGammaBurstsHelper(this);
+    return CanvasRenderer.drawGammaBursts.call(this);
   }
 
   /**
    * Paint the active radial tower command menu around the selected lattice.
    */
   drawTowerMenu() {
-    if (!this.ctx || !this.activeTowerMenu) {
-      return;
-    }
-    const tower = this.getActiveMenuTower();
-    if (!tower) {
-      return;
-    }
-    const geometry = this.getTowerMenuGeometry(tower);
-    if (!geometry) {
-      return;
-    }
-    const { options, optionRadius, ringRadius } = geometry;
-    const ctx = this.ctx;
-    ctx.save();
-
-    ctx.beginPath();
-    ctx.strokeStyle = 'rgba(139, 247, 255, 0.35)';
-    ctx.lineWidth = Math.max(1.2, optionRadius * 0.14);
-    ctx.arc(tower.x, tower.y, ringRadius, 0, Math.PI * 2);
-    ctx.stroke();
-
-    options.forEach((option) => {
-      const selected = Boolean(option.selected);
-      ctx.beginPath();
-      ctx.fillStyle = selected ? 'rgba(255, 228, 120, 0.32)' : 'rgba(12, 16, 26, 0.88)';
-      ctx.strokeStyle = selected ? 'rgba(255, 228, 120, 0.9)' : 'rgba(139, 247, 255, 0.75)';
-      ctx.lineWidth = Math.max(1.4, optionRadius * 0.16);
-      ctx.arc(option.center.x, option.center.y, optionRadius, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.stroke();
-
-      ctx.fillStyle = 'rgba(255, 255, 255, 0.92)';
-      ctx.font = `${Math.round(optionRadius * 0.95)}px "Cormorant Garamond", serif`;
-      ctx.textAlign = 'center';
-      ctx.textBaseline = 'middle';
-      ctx.fillText(option.icon || '?', option.center.x, option.center.y);
-    });
-
-    ctx.restore();
+    return CanvasRenderer.drawTowerMenu.call(this);
   }
 }
