+++ assets/playfield/constants.js
@@ -1,0 +1,2 @@
+export const PLAYFIELD_VIEW_DRAG_THRESHOLD = 6;
+export const PLAYFIELD_VIEW_PAN_MARGIN_METERS = 4;
